% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{bigleaf.surface}
\alias{bigleaf.surface}
\title{Big-leaf surface conditions}
\usage{
bigleaf.surface(data, Tair = "Tair", pressure = "pressure", LE = "LE",
  H = "H", Ca = "Ca", VPD = "VPD", Ga = "Ga", NEE = "NEE",
  constants = bigleaf.constants())
}
\arguments{
\item{data}{data.frame or matrix containing all required input variables}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Atmospheric pressure (kPa)}

\item{LE}{Latent heat flux (W m-2)}

\item{H}{Sensible heat flux (W m-2)}

\item{Ca}{Atmospheric CO2 concentration (mol mol-1)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{Ga}{Aerodynamic conductance (m s-1)}

\item{NEE}{Net ecosystem exchange (umol m-2 s-1)}

\item{constants}{}

\item{Rn}{Net radiation (W m-2)}
}
\value{
a data.frame with the following columns:
        \item{Tsurf}{Surface temperature (deg C)} \cr
        \item{esat_surf}{Saturation vapor pressure at the surface (kPa)} \cr
        \item{esurf}{vapor pressure at the surface (kPa)} \cr
        \item{VPD_surf}{vapor pressure deficit at the surface (kPa)} \cr
        \item{qsurf}{specific humidity at the surface} \cr
        \item{Ca_surf}{CO2 concentration at the surface}
}
\description{
Calculates meteorological conditions at the big-leaf surface
             by inverting bulk transfer equations for water, energy, and carbon
             fluxes.
}
\details{

}

