% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf_physiology.r
\name{bigleaf.Vcmax.Jmax}
\alias{bigleaf.Vcmax.Jmax}
\title{Bulk canopy biochemical parameters Vcmax and Jmax}
\usage{
bigleaf.Vcmax.Jmax(data, Temp, GPP, Ci, PPFD, PPFD_j = c(200, 500),
  PPFD_c = 1000, Oi = 0.21, Kc25 = 404.9, Ko25 = 278.4, Gam25 = 42.75,
  Kc_Ha = 79.43, Ko_Ha = 36.38, Gam_Ha = 37.83, Vcmax_Ha = 65.33,
  Vcmax_Hd = 200, Vcmax_dS = 635, Jmax_Ha = 43.9, Jmax_Hd = 200,
  Jmax_dS = 640, Theta = 0.7, alpha_canopy = 0.8,
  constants = bigleaf.constants())
}
\arguments{
\item{data}{Data.Frame or matrix with all required columns}

\item{Temp}{Surface (or air) temperature (degC)}

\item{GPP}{Gross primary productivty (umol m-2 s-1)}

\item{Ci}{Bulk canopy intercellular CO2 concentration (umol mol-1)}

\item{PPFD}{Photosynthetic photon flux density (umol m-2 s-1)}

\item{PPFD_j}{PPFD threshold, below which the canopy is considered to 
be RuBP regeneration limited. Defaults to 500 umol m-2 s-1.}

\item{PPFD_c}{PPFD threshold, above which the canopy is considered to 
be Rubisco limited. Defaults to 1000 umol m-2 s-1.}

\item{Oi}{Intercellular O2 concentration (mol mol-1)}

\item{Kc25}{Michaelis-Menten constant for CO2 at 25 degC (umol mol-1)}

\item{Ko25}{Michaelis-Menten constant for O2 at 25 degC (mmol mol-1)}

\item{Gam25}{Photorespiratory CO2 compensation point ('Gamma star') 
at 25 degC (umol mol-1)}

\item{Kc_Ha}{Activation energy for Kc (kJ mol-1)}

\item{Ko_Ha}{Activation energy for Ko (kJ mol-1)}

\item{Gam_Ha}{Activation energy for Gam (kJ mol-1)}

\item{Vcmax_Ha}{Activation energy for Vcmax (kJ mol-1)}

\item{Vcmax_Hd}{Deactivation energy for Vcmax (kJ mol-1)}

\item{Vcmax_dS}{Entropy term for Vcmax (J mol-1 K-1)}

\item{Jmax_Ha}{Activation energy for Jmax (kJ mol-1)}

\item{Jmax_Hd}{Deactivation energy for Jmax (kJ mol-1)}

\item{Jmax_dS}{Entropy term for Jmax (J mol-1 K-1)}

\item{Theta}{Curvature term in the light response function of J (-)}

\item{alpha_canopy}{Canopy absorptance (-)}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
Rgas - universal gas constant (J mol-1 K-1)}
}
\value{
a data.frame with the following columns:
        \item{Vcmax25}{maximum carboxylation rate at 25degC (umol m-2 s-1)}
        \item{Jmax25}{maximum electron transport rate at 25degC (umol m-2 s-1)}
}
\description{
Bulk canopy maximum carboxylation rate (Vcmax), electron transport rate (J),
             and maximum electron transport rate (Jmax) from bulk intercellular 
             CO2 concentration using the Farquhar et al. 1980 model for C3 photosynthesis.
}
\details{
The maximum carboxylation rate (Vcmax) and Vcmax25 (Vcmax at 25degC) are calculated
         as at leaf level. The required variables Gs and Ci can be calculated from 
         \code{\link{surface.conductance}} and \code{\link{intercellular.CO2}}, respectively.
         
         Gas exchange parameters are taken from Bernacchi et al. 2001 (assuming
         an infinite mesophyll conductance)
         
         Vcmax is calculated from the photosynthesis model by Farquhar et al. 1980,
         assuming that net photosynthesis is Rubisco-limited (RuBP-satured carboxylation
         rate, i.e. light has to be saturating):
         
         \deqn{Vcmax = (GPP * (Ci + Kc*(1.0 + Oi/Ko))) / Ci}
         
         Vcmax at canopy level is assumed to follow the same temperature response
         as at leaf level. Hence, Vcmax at 25degC (Vcmax25) is calculated as 
         (see e.g. Kattge & Knorr 2007):
         
         \deqn{Vcmax25 = Vcmax / 
                        ( exp(Vcmax_Ha * (Temp - Tref) / (Tref*constants$Rgas*Temp)) *
                        (1 + exp((Tref*dS - Vcmax_Hd) / (Tref * constants$Rgas))) /
                        (1 + exp((Temp*dS - Vcmax_Hd) / (Temp * constants$Rgas)))
                        )
               }
}
\note{
The critical assumption is that bulk canopy photosynthesis is Rubisco-
        limited. Therefore, only GPP data above a certain light threshold are
        used for the calculation of Vcmax. This threshold (\code{PPFD_c}) is
        likely dependent on the canopy of the ecosystem and presumably higher
        for dense than for open canopies. A threshold of 500 umol m-2 s-1 PPFD
        is a reasonable working assumption (see Kosugi et al. 2013).
        The photosynthesis parameters are also likely dependent on the ecosystem
        and growth conditions.
}
\examples{
carbox.rate(Temp=20,GPP=15,Ci=300,PPFD=1000,PPFD_c=500)  
                                      
}
\references{
Lloyd J. et al., 1995: A simple calibrated model of Amazon rainforest productivity
            based on leaf biochemical properties. Plant, Cell and Environment 18, 1129-1145.

            Rayment M.B., Loustau D., Jarvis P.G., 2002: Photosynthesis and respiration
            of black spruce at three organizational scales: shoot, branch and canopy.
            Tree Physiology 22, 219-229.

            Kosugi Y. et al., 2013: Determination of the gas exchange phenology in an
            evergreen coniferous forest from 7 years of eddy covariance flux data using
            an extended big-leaf analysis. Ecol Res 28, 373-385. 
            
            Ueyama M. et al, 2016: Optimization of a biochemical model with eddy covariance
            measurements in black spruce forests of Alaska for estimating CO2 fertilization
            effects. Agricultural and Forest Meteorology 222, 98-111.
            
            Bernacchi C.J., Singsaas E.L., Pimentel C., Portis JR A.R., Long S.P., 2001:
            Improved temperature response functions for models of Rubisco-limited
            photosynthesis. Plant, Cell and Environment 24, 253-259. 
            
            Bernacchi C.J., Pimentel C., Long S.P., 2003: In vivo temperature response
            functions of parameters required to model RuBP-limited photosynthesis.
            Plant, Cell and Environment 26, 1419-1430.
            
            Kattge J., Knorr W., 2007: Temperature acclimation in a biochemical
            model of photosynthesis: a reanalysis of data from 36 species.
            Plant, Cell and Environment 30, 1176-1190.
}
