% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Ca.surface}
\alias{Ca.surface}
\title{Canopy surface CO2 concentration}
\usage{
Ca.surface(Ca, NEE, Ga_CO2)
}
\arguments{
\item{Ca}{atmospheric CO2 concentration (umol mol-1)}

\item{NEE}{net ecosystem exchange (umol CO2 m-2 s-1)}

\item{Ga_CO2}{aerodynamic conductance for CO2 (mol m-2 s-1)}
}
\value{
Ca_surf CO2 concentration at the canopy surface (umol mol-1)
}
\description{
the CO2 concentration at the canopy surface derived from net ecosystem
             CO2 exchange and measured atmospheric CO2 concentration.
}
\note{
the following sign convention is employed: negative values of NEE denote
      net CO2 uptake by the ecosystem.
}

