% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{carbox.rate}
\alias{carbox.rate}
\title{Bulk canopy maximum carboxylation efficiency (Vcmax)}
\usage{
carbox.rate(Temp, GPP, Ci, PPFD, PPFD_c, Oi = 0.21, Kc25 = 404.9,
  Ko25 = 278.4, Kc_Ha = 79.43, Ko_Ha = 36.38, Vcmax_Ha = 65.33,
  Vcmax_Hd = 200, dS = 640, constants = bigleaf.constants())
}
\arguments{
\item{Temp}{Surface (or air) temperature (degC)}

\item{GPP}{Gross primary productivty (umol m-2 s-1)}

\item{Ci}{Bulk canopy intercellular CO2 concentration (umol mol-1)}

\item{PPFD}{Photosynthetic photon flux density (umol m-2 s-1)}

\item{PPFD_c}{PPFD threshold, above which the canopy is considered to 
be light saturated (and Rubisco limited).}

\item{Oi}{intercellular O2 concentration (mol mol-1)}

\item{Kc25}{Michaelis-Menten constant for CO2 at 25 degC (umol mol-1)}

\item{Ko25}{Michaelis-Menten constant for O2 at 25 degC (mmol mol-1)}

\item{Kc_Ha}{activation energy for Kc (kJ mol-1)}

\item{Ko_Ha}{activation energy for Ko (kJ mol-1)}

\item{Vcmax_Ha}{activation energy for Vcmax (kJ mol-1)}

\item{Vcmax_Hd}{deactivation energy for Vcmax (kJ mol-1)}

\item{dS}{entropy term (J mol-1 K-1)}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
Rgas - universal gas constant (J mol-1 K-1)}
}
\value{
a data.frame with the following columns:
        \item{Vcmax}{maximum carboxylation rate at Temp (umol m-2 s-1)}
        \item{Vcmax25}{maximum carboxylation rate at 25degC (umol m-2 s-1)}
        \item{Ko}{Michaelis-Menten constant for O2 at Temp (mmol mol-1)}
        \item{Kc}{Michaelis-Menten constant for CO2 at Temp (umol mol-1)}
}
\description{
Bulk canopy maximum carboxylation rate (Vcmax) from bulk intercellular 
             CO2 concentration using the Farquhar et al. 1980 model for C3 photosynthesis.
}
\details{
The maximum carboxylation rate (Vcmax) and Vcmax25 (Vcmax at 25degC) are calculated
         as at leaf level. The required variables Gs and Ci can be calculated from 
         \code{\link{surface.conductance}} and \code{\link{intercellular.CO2}}, respectively.
         
         Gas exchange parameters are taken from Bernacchi et al. 2001 (assuming
         an infinite mesophyll conductance)
         
         Vcmax is calculated from the photosynthesis model by Farquhar et al. 1980,
         assuming that net photosynthesis is Rubisco-limited (RuBP-satured carboxylation
         rate, i.e. light has to be saturating):
         
         \deqn{Vcmax = (GPP * (Ci + Kc*(1.0 + Oi/Ko))) / Ci}
         
         Vcmax at canopy level is assumed to follow the same temperature response
         as at leaf level. Hence, Vcmax at 25degC (Vcmax25) is calculated as 
         (see e.g. Kattge & Knorr 2007):
         
         \deqn{Vcmax25 = Vcmax / 
                        ( exp(Vcmax_Ha * (Temp - Tref) / (Tref*constants$Rgas*Temp)) *
                        (1 + exp((Tref*dS - Vcmax_Hd) / (Tref * constants$Rgas))) /
                        (1 + exp((Temp*dS - Vcmax_Hd) / (Temp * constants$Rgas)))
                        )
               }
}
\note{
The critical assumption is that bulk canopy photosynthesis is Rubisco-
        limited. Therefore, only GPP data above a certain light threshold are
        used for the calculation of Vcmax. This threshold (\code{PPFD_c}) is
        likely dependent on the canopy of the ecosystem and presumably higher
        for dense than for open canopies. A threshold of 500 umol m-2 s-1 PPFD
        is a reasonable working assumption (see Kosugi et al. 2013).
        The photosynthesis parameters are also likely dependent on the ecosystem
        and growth conditions.
}
\examples{
carbox.rate(Temp=20,GPP=15,Ci=300,PPFD=1000,PPFD_c=500)  
                                      
}
\references{
Kosugi Y. et al., 2013: Determination of the gas exchange phenology in an
            evergreen coniferous forest from 7 years of eddy covariance flux data using
            an extended big-leaf analysis. Ecol Res 28, 373-385. 
            
            Bernacchi C.J, Singsaas E.L., Pimentel C., Portis JR A.R., Long S.P., 2001:
            Improved temperature response functions for models of Rubisco-limited
            photosynthesis. Plant, Cell and Environment 24, 253-259. 
            
            Kattge J., Knorr W., 2007: Temperature acclimation in a biochemical
            model of photosynthesis: a reanalysis of data from 36 species.
            Plant, Cell and Environment 30, 1176-1190.
}
