% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.r
\name{check.columns}
\alias{check.columns}
\title{checks columns in a data frame}
\usage{
check.columns(data, varname)
}
\arguments{
\item{data}{Input data.frame or a matrix}

\item{varname}{Input variable. Must be a character}
}
\description{
check if columns in a data.frame or matrix (or vector variables)
             exist and if they are of the right type (numeric).
}
\details{
The Input variable 'varname' must be of type character. The function
         searches for the object 'varname' in the parent environment and returns it.
         The function then ensures that the requirements of the input variable (E.g. does
         the column exist? Is the input numeric? etc.) is met.
}
