% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.r
\name{check.input}
\alias{check.input}
\title{checks input for functions in the bigleaf package}
\usage{
check.input(data, ...)
}
\arguments{
\item{data}{Input data.frame or matrix}

\item{...}{Input variables. Either a list or individual vectors}
}
\description{
Checks length and type of the provided input variables.
}
