% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.r
\name{check.length}
\alias{check.length}
\title{Test Variables for Equal Length}
\usage{
check.length(varlist)
}
\arguments{
\item{varlist}{List of variables for which the length has to be compared}
}
\description{
Test Variables for Equal Length
}
\note{
This function only plays a role if no input data.frame or matrix are 
      provided. In this case it ensures that provided vectors have the same
      length to avoid trouble later up the function call.
}
\keyword{internal}
