% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_radiation.r
\name{computePotentialRadiation}
\alias{computePotentialRadiation}
\title{potential radiation}
\usage{
computePotentialRadiation(doy, hour, latDeg, longDeg, timezone,
  useSolartime = TRUE)
}
\arguments{
\item{doy}{integer vector with day of year (start at 1),
same length as hour or length 1}

\item{hour}{numeric vector with daytime as decimal hour of local time zone}

\item{latDeg}{Latitude in (decimal) degrees}

\item{longDeg}{Longitude in (decimal) degrees}

\item{timezone}{Time zone (in hours)}

\item{useSolartime}{by default corrects hour (given in local winter time)
for latitude to solar time (where noon is exactly at 12:00).
Set this to FALSE to directly use local winter time}
}
\value{
vector of potential radiation (PotRad, W_m-2)
}
\description{
Compute potential radiation for given geolocation and day of yeat
}
\examples{
hour <- seq(5, 18, by = 0.1)
potRadApparentLocal <- computePotentialRadiation(
  160, hour, 39.94, -5.77, timezone = +1)
potRadTimezone <- computePotentialRadiation(
  160, hour, 39.94, -5.77, timezone = +1, useSolartime = FALSE)
plot(potRadApparentLocal ~ hour, type = 'l'
  , ylab = 'potential radiation (W m-2)')
lines(potRadTimezone ~  hour, col = "blue")
abline(v = 12, col = "blue", lty = "dotted")
legend("bottomright", legend = c("solar time", "local winter time")
, col = c("black", "blue"), inset = 0.05, lty = 1)
}
