% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorological_variables.r
\name{dew.point.solver}
\alias{dew.point.solver}
\title{Solver function for dew point temperature}
\usage{
dew.point.solver(ea, accuracy)
}
\arguments{
\item{ea}{Air vapor pressure (kPa)}

\item{accuracy}{Accuracy of the result (degC)}
}
\description{
Solver function used in dew.point()
}
\note{
\code{accuracy} is passed to this function by dew.point().
}
\keyword{internal}
