% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{e.to.q}
\alias{VPD.to.e}
\alias{VPD.to.q}
\alias{VPD.to.rH}
\alias{e.to.VPD}
\alias{e.to.q}
\alias{q.to.VPD}
\alias{q.to.e}
\alias{rH.to.VPD}
\title{Conversions between humidity measures}
\usage{
e.to.q(e, pressure, constants = bigleaf.constants())

q.to.e(q, pressure, constants = bigleaf.constants())

e.to.VPD(e, pressure, Tair)

VPD.to.e(VPD, pressure, Tair)

rH.to.VPD(rH, Tair)

VPD.to.rH(VPD, Tair)

q.to.VPD(q, Tair, pressure, constants = bigleaf.constants())

VPD.to.q(VPD, Tair, pressure, constants = bigleaf.constants())
}
\arguments{
\item{e}{Vapor pressure (kPa)}

\item{constants}{eps - ratio of the molecular weight of water vapor to dry air (-)}

\item{q}{Specific humidity (kg kg-1)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{rH}{Relative humidity (-)}

\item{p}{Air pressure (kPa)}
}
\description{
Conversion between vapor pressure, vapor pressure deficit (VPD), specific humidity, and relative humidity.
}
\references{
Foken, T, 2008: Micrometeorology. Springer, Berlin, Germany.
}

