% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Esat}
\alias{Esat}
\title{Saturation vapor pressure (Esat) and slope of the Esat curve}
\usage{
Esat(Tair, formula = c("Sonntag_1990", "Alduchov_1996"))
}
\arguments{
\item{Tair}{Air temperature (deg C)}

\item{formula}{Formula to be used. Either "Sonntag_1990" (Default) or "Alduchov_1996"}
}
\value{
A dataframe with the following columns: 
   \item{Esat}{Saturation vapor pressure (kPa)}
   \item{Delta}{Slope of the saturation vapor pressure curve (kPa K-1)}
}
\description{
Calculates Saturation vapor pressure (Esat) over water and the
             corresponding slope of the saturation vapor pressure curve.
}
\details{
Esat (kPa) is calculated based on the Magnus equation:

 \deqn{Esat = a * exp((b * Tair) / (c + Tair)) / 1000}
 
 where the coefficients a, b, c take different values depending on the formula used.
 The slope of the Esat curve (\eqn{\Delta}) is calculated as the first derivative of the function:
 
 formula
}
\examples{
Esat(seq(0,45,5))[,"Esat"]  # Esat in kPa
Esat(seq(0,45,5))[,"Delta"] # the corresponding slope of the Esat curve in kPa K-1
            
}
\references{
Sonntag_1990 Important new values of the physical constants of 1986, vapor 
            pressure formulations based on the ITS-90, and psychrometric formulae 
            
            Alduchov, O. A. & Eskridge, R. E., 1996: Improved Magnus form approximation of 
            saturation vapor pressure. Journal of Applied Meteorology, 1996, 35, 601-609
}
