% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{ET.components}
\alias{ET.components}
\title{Imposed and Equilibrium Evapotranspiration}
\usage{
\method{ET}{components}(data, Tair = "Tair", pressure = "pressure",
  VPD = "VPD", Rn = "Rn", constants = bigleaf.constants())
}
\arguments{
\item{data}{data.frame or matrix containing all required input variables}

\item{Tair}{air temperature (deg C)}

\item{pressure}{air pressure (kPa)}

\item{VPD}{air vapor pressure deficit (kPa)}

\item{Rn}{net radiation}

\item{constants}{}

\item{Gs}{surface conductance (m s-1)}
}
\value{

}
\description{
Imposed and Equilibrium Evapotranspiration
}
\details{
Evapotranspiration (ET) split up in imposed ET and equilibrium ET.
}
\references{
Jarvis P.G., McNaughton K.G., 1986: Stomatal control of transpiration:
            scaling up from leaf to region. Advances in Ecological Research 15, 1-49.
            
            Monteith J.L., Unsworth M.H., 2008: Principles of environmental physics.
            3rd edition. Academic Press, London.
}

