% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{ET.pot}
\alias{ET.pot}
\title{Potential evapotranspiration from the Priestley-Taylor equation}
\usage{
ET.pot(data, Tair = "Tair", pressure = "pressure", Rn = "Rn", G = NULL,
  S = NULL, alpha = 1.26, missing.G.as.NA = F, missing.S.as.NA = F,
  constants = bigleaf.constants())
}
\arguments{
\item{data}{Data.frame or matrix containing all required variables}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{Rn}{Net radiation (W m-2)}

\item{G}{Ground heat flux (W m-2)}

\item{S}{Sum of all storage fluxes (W m-2)}

\item{alpha}{Priestley-Taylor coefficient (-)}

\item{constants}{cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
eps - ratio of the molecular weight of water vapor to dry air (-)}
}
\value{
a data.frame with the following columns:
        \item{ET_pot}{potential evapotranspiration (kg m-2 s-1)}
        \item{LE_pot}{potential latent heat flux (W m-2)}
}
\description{
Potential evapotranspiration from the Priestley-Taylor equation
}
\details{


Potential evapotranspiration is calculated according to Priestley & Taylor, 1972:

         \deqn{(\alpha * \delta * (Rn - G - S)) / (\delta + \gamma)}
}
\references{
Priestley C.H.B., Taylor R.J., 1972: On the assessment of surface heat flux
            and evaporation using large-scale parameters. Monthly Weather Review 100, 81-92.
}

