% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{ET.ref}
\alias{ET.ref}
\title{Reference evapotranspiration}
\usage{
ET.ref(data, Gs, Tair = "Tair", pressure = "pressure", VPD = "VPD",
  Ga = "Ga", Rn = "Rn", G = NULL, S = NULL, missing.G.as.NA = F,
  missing.S.as.NA = F, constants = bigleaf.constants())
}
\arguments{
\item{data}{Data.frame or matrix containing all required variables}

\item{Gs}{Surface conductance (m s-1)}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{Ga}{Aerodynamic conductance (m s-1)}

\item{Rn}{Net radiation (W m-2)}

\item{G}{Ground heat flux (W m-2); optional}

\item{S}{Sum of all storage fluxes (W m-2); optional}

\item{missing.G.as.NA}{if TRUE, missing G are treated as NA,otherwise set to 0.}

\item{missing.S.as.NA}{if TRUE, missing S are treated as NA,otherwise set to 0.}

\item{constants}{cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
eps - ratio of the molecular weight of water vapor to dry air (-) \cr
Rd - gas constant of dry air (J kg-1 K-1) \cr
Rgas - universal gas constant (J mol-1 K-1) \cr
Kelvin - conversion degree Celsius to Kelvin \cr}
}
\value{
a data.frame with the following columns:
        \item{ET_0}{reference evapotranspiration (kg m-2 s-1)}
        \item{LE_0}{reference latent heat flux (W m-2)}
}
\description{
Reference evapotranspiration
}
\details{
Reference evapotranspiration calculated from the Penman-Monteith
         equation with a pre-defined surface conductance.

Reference evapotranspiration is calculated according to the Penman-Monteith
         equation:

         \deqn{LE_0 = (\delta * (Rn - G - S) * \rho * cp * VPD * Ga) / (\delta + \gamma * (1 + Ga/Gs)}
}
\references{

}
