% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{filter.growseas}
\alias{filter.growseas}
\title{GPP-based growing season filter}
\usage{
filter.growseas(GPPd, tGPP, min_int, ws)
}
\arguments{
\item{GPPd}{daily GPP}

\item{tGPP}{GPP threshold (fraction of 95% quantile)}

\item{ws}{window size used for smoothing}
}
\description{
filters annual time series for growing season based on smoothed daily GPP time series
}

