% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.r
\name{g.to.umol}
\alias{g.to.umol}
\title{convert mass in g to amount of substance in umol}
\usage{
g.to.umol(massFlux, molarMass, constants = bigleaf.constants())
}
\arguments{
\item{massFlux}{numeric vector of mass in g}

\item{molarMass}{numeric vector of molar mass of the substance}

\item{constants}{list with entry kg2g, see \code{\link{bigleaf.constants}}}
}
\value{
numeric vector of amoutn of substance in umol
}
\description{
convert mass in g to amount of substance in umol
}
