% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Ga}
\alias{Ga}
\title{Aerodynamic conductance}
\usage{
Ga(data, Tair = "Tair", pressure = "pressure", wind = "wind",
  ustar = "ustar", H = "H", zr, zh, d, z0m, frac_d = 0.7,
  frac_z0m = 0.1, Dl, N, LAI, fc = NULL, Cd = 0.2, hs = 0.01,
  stab_correction = c("Dyer_1970", "Businger_1971", "none"),
  method_roughness = c("canopy_height", "wind_profile"),
  Rb_model = c("Thom_1972", "Choudhury_1988", "Su_2001", "constant_kB"),
  kB = NULL, constants = bigleaf.constants())
}
\arguments{
\item{Tair}{air temperature (degC)}

\item{pressure}{air pressure (Pa)}

\item{wind}{wind speed (m s-1)}

\item{ustar}{friction velocity (m s-1)}

\item{H}{sensible heat flux (W m-2)}

\item{zr}{measurement (=reference) height (m)}

\item{zh}{canopy height (m)}

\item{d}{zero-plane displacement height (m)}

\item{z0m}{roughness length for momentum (m)}

\item{Dl}{characteristic leaf dimension (m)}

\item{N}{number of leaf sides participating in heat exchange (1 or 2)}

\item{LAI}{one-sided leaf area index (m2 m-2)}

\item{fc}{fractional vegetation cover (-)}

\item{Cd}{foliage drag coefficient (-)}

\item{hs}{roughness length of bare soil (m)}

\item{stab_correction}{stability correction function}

\item{Rb_model}{boundary layer resistance formulation}

\item{kB}{kB-1 value; only used if \code{Rb_model = "constant_kB"}}
}
\value{
a dataframe with the following columns:
}
\description{
Aerodynamic conductance, including options for the boundary layer conductance
             formulation and stability correction functions.
}
\details{
Aerodynamic conductance (Ga) is calculated as:

 Ga = 1/Ra_m + 1/Rb
 
 where Ra_m is the aerodynamic resistance for momentum and Rb the canopy boundary
 layer resistance.
 
 The argument stab_correction determines the stability correction function used 
 to account for the effect of atmospheric stability on Ra_m (Ra_m is lower for unstable
 and higher for stable stratification). Stratification is based on a stability parameter zeta (z-d/L),
 where z = reference height, d the zero-plane displacement height, and L the Monin-Obukhov length, 
 calculated with \code{\link{MoninObukhov.length}}
 Stability correction functions calculate the correction based on the formulations by Businger 1971 
 Dyer 1970, based on zeta.
 If "none", atmospheric stability is neglected, and Monin-Obukhov length, as well as the 
 stability parameter zeta are not calculated.

 The argument \code{Rb_model} determines the canopy boundary layer resistance ("excess resistance") model that is used. "Thom_1972" is an 
 empirical formulation based on the friction velocity (ustar) (Thom 1972):
 
   \deqn{Rb = 0.667u*^0.667}
   
 The model by Mcnaughton and van den Hurk 1995 ("McNaughton_1995"), calculates Rb
 based on leaf width, LAI and ustar (Note that the original formulation leaf width
 instead of the characteristic leaf dimension (Dl) is taken):
  
    formula
    
 The option "Su_2001" calculates Rb based on the physically-based Rb model by Su et tal. 2001,
 a simplification of the model developed by Massmann 1999:
 
    formula
 
 The models calculate the ... parameter kB-1, which related to Rb:
 
    formula
 
 Both kB and Rb are given in the output.
}
\references{

}

