% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Gb.McNaughton}
\alias{Gb.McNaughton}
\title{Boundary layer conductance according to McNaughton and van den Hurk 1995 --> check units in original paper!}
\usage{
Gb.McNaughton(ustar, leafwidth, LAI, constants)
}
\arguments{
\item{ustar}{friction velocity (m s-1)}

\item{leafwidth}{leaf width (m)}

\item{LAI}{one-sided leaf area index}

\item{k}{von-Karman constant (-)}
}
\description{
Boundary layer conductance according to McNaughton and van den Hurk 1995 --> check units in original paper!
}
\seealso{
\code{\link{Gb.Thom}}, \code{\link{Gb.Su}}
}

