% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Gr.longwave}
\alias{Gr.longwave}
\title{longwave radiative transfer conductance of the canopy}
\usage{
Gr.longwave(Tair, LAI, constants = bigleaf.constants())
}
\arguments{
\item{Tair}{Air temperature (deg C)}

\item{LAI}{Leaf area index (m2 m-2)}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
sigma - Stefan-Boltzmann constant (W m-2 K-4) \cr
cp - specific heat of air for constant pressure (J K-1 kg-1)}
}
\value{
Gr longwave radiative transfer conductance of the canopy (m s-1)
}
\description{
longwave radiative transfer conductance of the canopy
}
\details{
the following formula is used (Martin, 1989):

         \deqn{Gr = 4 \sigma Tair^3 LAI / cp}
}
\references{
Martin P., 1989: The significance of radiative coupling between
            vegetation and the atmosphere. Agricultural and Forest Meteorology 49, 45-53.
}

