% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{GsPM}
\alias{GsPM}
\title{Surface conductance for water}
\usage{
GsPM(Tair, pressure, Rn, VPD, LE, Ga, constants)
}
\arguments{
\item{Tair}{Air temperature (deg C)}

\item{pressure}{Atmospheric pressure (kPa)}

\item{Rn}{Net radiation (W m-2)}

\item{VPD}{Vapor pressure deficit (Pa)}

\item{LE}{Latent heat flux (W m-2)}

\item{Ga}{Aerodynamic conductance (m s-1)}

\item{G}{Ground heat flux (W m-2)}

\item{S}{Sum of all storage fluxes (W m-2)}
}
\value{
a dataframe with two columns: 

 Gs_ms:  Surface conductance in ms-1
 Gs_mol: Surface conductance in mol m-2 s-1
}
\description{
Calculates surface conductance for water from the inverted Penman-Monteith
             equation.
}
\details{
Surface conductance is calculated from the inverted Penman-Monteith equation:

 formula
 
 Note that Gs > Gc (canopy conductance) in time periods when a significant fraction of 
 ET comes from interception or soil evaporation. 
 Available energy A is given by (Rn - G - S). If G and/or S are not provided, then A = Rn.
}

