% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{hypsometric.equation}
\alias{hypsometric.equation}
\title{Air pressure from hypsometric equation}
\usage{
hypsometric.equation(elev, Tair, q = NULL, constants = bigleaf.constants())
}
\arguments{
\item{elev}{elevation a.s.l. (m)}

\item{Tair}{air temperature (degC)}

\item{q}{specific humidity (kg kg-1); optional}

\item{constants}{Kelvin- conversion degC to Kelvin \cr
pressure0 - reference atmospheric pressure at sea level (Pa) \cr
Rd - gas constant of dry air (J kg-1 K-1) \cr
g -  gravitational acceleration (m s-2) \cr}
}
\value{
air pressure (kPa)
}
\description{
An estimate of mean pressure at a given elevation as predicted by the
             hypsometric equation
}
\details{

}
\note{
the hypsometric equation gives a rough estimate of the standard pressure
      at a given altitude. If specific humidity q is provided, humidity correction
      is applied and the virtual temperature instead of air temperature is used.
}
\references{
Stull, B., 1988: An Introduction to Boundary Layer Meteorology.
            Kluwer Academic Publishers, Dordrecht, Netherlands
}

