% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.r
\name{kg.to.mol}
\alias{kg.to.mol}
\title{convert mass in g to amount of substance in umol}
\usage{
kg.to.mol(mass, molarMass = bigleaf.constants()$H2Omol)
}
\arguments{
\item{mass}{numeric vector of mass in kg}

\item{molarMass}{numeric vector of molar mass of the substance in kg/mol
e.g. as provided by \code{\link{bigleaf.constants}}()$H2Omol
Default is molar mass of Water.}
}
\value{
numeric vector of amount of substance in mol
}
\description{
convert mass in g to amount of substance in umol
}
