% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{LE.vaporization}
\alias{LE.vaporization}
\title{Latent heat of vaporization}
\usage{
LE.vaporization(Tair)
}
\arguments{
\item{Tair}{Air temperature (deg C)}
}
\value{
Latent heat of vaporization \eqn{\lambda} (J kg-1)
}
\description{
Latent heat of vaporization as a function of air temperature.
}
\details{
The following formula is used:

\deqn{\lambda = (2.501 - 0.00237*Tair)10^6}
}
\examples{
LE.vaporization(seq(5,45,5))             
            
}
\references{
Stull, B., 1988: An Introduction to Boundary Layer Meteorology (p.641)
            Kluwer Academic Publishers, Dordrecht, Netherlands
}
