% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{LEtoET}
\alias{ETtoLE}
\alias{LEtoET}
\title{Conversion between latent heat flux and evapotranspiration}
\usage{
LEtoET(LE, Tair)
}
\arguments{
\item{LE}{Latent heat flux (W m-2)}

\item{Tair}{Air temperature (deg C)}

\item{ET}{Evapotranspiration (kg m-2 s-1)}
}
\description{
converts evpaorative water flux from mass (ET=evapotranspiration)
             to energy (LE=latent heat) units, or vice versa.
}

