% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{MoninObukhov.length}
\alias{MoninObukhov.length}
\title{Monin-Obukhov length}
\usage{
MoninObukhov.length(Tair, pressure, ustar, H, constants = bigleaf.constants())
}
\arguments{
\item{Tair}{air temperature (degC)}

\item{pressure}{air pressure (kPa)}

\item{ustar}{friction velocity (m s-1)}

\item{H}{sensible heat flux (W m-2)}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
k - von Karman constant (-) \cr
g - gravitational acceleration (m s-2)}
}
\value{
Monin-Obukhov length (m)
}
\description{


The Monin-Obukhov length (L) is given by:

             \deqn{L = - (\rho * cp * ustar^3 * Tair) / (k * g * H)}
}
\references{
Foken, T, 2008: Micrometeorology. Springer, Berlin, Germany.
}

