% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{mstomol}
\alias{moltoms}
\alias{mstomol}
\title{Conversion between conductance units}
\usage{
mstomol(Gc_ms, Tair, pressure, constants = bigleaf.constants())

moltoms(Gc_mol, Tair, pressure, constants = bigleaf.constants())
}
\arguments{
\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{constants}{}

\item{Gs_ms}{Canopy/Surface conductance (ms-1)}

\item{Gs_mol}{Canopy/Surface conductance (mol m-2 s-1)}
}
\description{
Converts canopy/surface conductance from ms-1
             to mol m-2 s-1, or vice versa
}
\references{
Jones, H.G. 1992. Plants and microclimate: a quantitative approach to environmental plant physiology.
            2nd Edition., 2nd Edn. Cambridge University Press, Cambridge. 428 p --> replace with third edition
}

