% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Photosyn_energy}
\alias{Photosyn_energy}
\title{biochemical energy (Heat)}
\usage{
Photosyn_energy(alpha = 0.422, NEE)
}
\arguments{
\item{alpha}{energy taken up/released by photosynthesis/respiration (J umol-1)}

\item{NEE}{net ecosystem exchange (umol CO2 m-2 s-1)}
}
\value{
biochemical energy Sp (W m-2)
}
\description{
radiant energy absorbed in photosynthesis or heat release by respiration based on NEE
}
\details{
the following sign convention is employed: NEE is negative when carbon is taken up by the ecosystem.
         Positive values in the result mean that energy is taken up by the ecosystem, negative ones that heat is 
         released.
         The value of alpha is taken from Nobel_1974 (Meyers_2004), but other values exist (Blanken_1997)
}

