% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{psychrometric.constant}
\alias{psychrometric.constant}
\title{Psychrometric constant}
\usage{
psychrometric.constant(Tair, pressure, constants)
}
\arguments{
\item{Tair}{air temperature (degC)}

\item{pressure}{air pressure (Pa)}
}
\value{
the psychrometric constant (Pa K-1)
}
\description{
Calculates the psychrometric 'constant'
}
\details{
The psychrometric constant is given as:

 formula
}
\references{
test
}

