% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{ReynoldsNumber}
\alias{ReynoldsNumber}
\title{Roughness Reynolds Number}
\usage{
ReynoldsNumber(hs, ustar, pressure, Tair, constants)
}
\arguments{
\item{ustar}{friction velocity (m s^{-1})}

\item{k}{von-Karman constant (-)}
}
\value{
a matrix containing Gb, Rb, and kB
}
\description{
An empirical formulation for the quasi-laminar boundary layer conductance
based on ustar
}

