% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{stability.correction.H}
\alias{stability.correction.H}
\title{stability correction functions for sensible heat}
\usage{
stability.correction.H(Tair, pressure, ustar, H, zr, d,
  formulation = c("Dyer_1970", "Businger_1971"),
  constants = bigleaf.constants())
}
\arguments{
\item{formulation}{formulation for the stability function. Either "Dyer_1970", or "Businger_1971"}

\item{zeta}{stability parameter \eqn{\zeta}}
}
\value{
psi_h the value of the stability function for sensible heat (-)
}
\description{
dimensionless stability functions needed to correct deviations
             from the exponential wind profile under non-neutral conditions.
}
\details{
The functions depend on the value of the stability parameter \eqn{\zeta},
         which can be calculated from the function \code{\link{stability.parameter}}.
         The general form of the stability functions is:
         
         \deqn{-x * zeta} 
         
         for stable atmospheric conditions (\eqn{\zeta} >= 0), and
         
         \deqn{2 * log( (1 + y) / 2) }
         
         for unstable atmospheric conditions (\eqn{\zeta} < 0).
         
         The different formulations differ in their value of x and y.
}
\references{
Dyer, A.J., 1974: A review of flux-profile relationships. 
            Boundary-Layer Meteorology 7, 363-372.
            
            Dyer, A. J., Hicks, B.B., 1970: Flux-Gradient relationships in the
            constant flux layer. Quart. J. R. Meteorol. Soc. 96, 715-721.
            
            Businger, J.A., Wyngaard, J. C., Izumi, I., Bradley, E. F., 1971:
            Flux-Profile relationships in the atmospheric surface layer. 
            J. Atmospheric Sci. 28, 181-189.

            Foken, T, 2008: Micrometeorology. Springer, Berlin, Germany.
}

