% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf_physiology.r
\name{stomatal.sensitivity}
\alias{stomatal.sensitivity}
\title{Stomatal sensitivity to VPD}
\usage{
stomatal.sensitivity(data, Gs = "Gs", VPD = "VPD", ...)
}
\arguments{
\item{data}{Data.frame or matrix containing all required columns}

\item{Gs}{Surface conductance (mol m-2 s-1)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{...}{Additional arguments to \code{\link[stats]{nls}}}
}
\value{
An nls model object containing (amongst others) estimates for the mean
        and standard errors of the parameters m and b.
}
\description{
Sensitivity of surface conductance to vapor pressure deficit.
}
\details{
The function fits the following equation (Oren et al. 1999):

         \deqn{Gs = -m ln(VPD) + b}

         where b is the reference surface conductance (Gs) at VPD=1kPa (in mol m-2 s-1),
         and m is the sensitvity parameter of Gs to VPD (in mol m-2 s-1 log(kPa)-1).
         The two parameters b and m are fitted using \code{\link[stats]{nls}}.
         VPD can be the one directly measured at instrument height, or the
         one at the surface, as returned by \code{\link{surface.conditions}}.
}
\references{
Oren R., et al. 1999: Survey and synthesis of intra- and interspecific
            variation in stomatal sensitivity to vapour pressure deficit. Plant,
            Cell & Environment 22, 1515-1526. 
            
            Novick K.A., et al. 2016: The increasing importance of atmospheric demand
            for ecosystem water and carbon fluxes. Nature Climate Change 6, 1023 - 1027.
}
