% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Trad.surface}
\alias{Trad.surface}
\title{Radiometric surface temperature}
\usage{
Trad.surface(longwave.up, emissivity, constants = bigleaf.constants())
}
\arguments{
\item{longwave.up}{longwave upward radiation (W m-2)}

\item{emissivity}{infrared emissivity of the surface (-)}

\item{constants}{sigma - Stefan-Boltzmann constant (W m-2 K-4) \cr
Kelvin - conversion degree Celsius to Kelvin}
}
\value{
Radiometric surface temperature in Kelvin and degC.
}
\description{
Radiometric surface temperature from longwave upward radiation
             measurements.
}
\details{
Radiometric surface temperature (Trad) is calculated as:

         \deqn{Trad = LW_up / (\sigma \epsilon)^(1/4)}
}

