% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{umolCO2.to.gC}
\alias{umolCO2.to.gC}
\alias{gC.to.umolCO2}
\title{Conversion between mass and molar units}
\usage{
umolCO2.to.gC(CO2_flux, constants = bigleaf.constants())

gC.to.umolCO2(C_flux, constants = bigleaf.constants())
}
\arguments{
\item{CO2_flux}{CO2 flux (umol CO2 m-2 s-1)}

\item{constants}{Cmol - molar mass of carbon (kg mol-1)}

\item{C_flux}{C flux (gC m-2 s-1)}

\item{C_flux}{CO2 flux (gC m-2 s-1)}
}
\description{
Converts CO2 quantities from umol CO2 m-2 s-1 to gC and vice versa.
}
\details{
The conversions are given by:
}
