% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorological_variables.r
\name{wetbulb.solver}
\alias{wetbulb.solver}
\title{Solver Function for Wet-Bulb Temperature}
\usage{
wetbulb.solver(ea, Tair, gamma, accuracy)
}
\arguments{
\item{ea}{Air vapor pressure (kPa)}

\item{Tair}{Air temperature (degC)}

\item{gamma}{Psychrometric constant (kPa K-1)}

\item{accuracy}{Accuracy of the result (degC)}
}
\description{
Solver function used in wetbulb.temp()
}
\note{
\code{accuracy} is passed to this function by wetbulb.temp().
}
\keyword{internal}
