% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{aerodynamic.conductance}
\alias{aerodynamic.conductance}
\title{Aerodynamic conductance}
\usage{
aerodynamic.conductance(data, Tair = "Tair", pressure = "pressure",
  wind = "wind", ustar = "ustar", H = "H", zr, zh, d, z0m, Dl, N,
  fc = NULL, LAI, Cd = 0.2, hs = 0.01, stab_correction = TRUE,
  stab_formulation = c("Dyer_1970", "Businger_1971"),
  Rb_model = c("Thom_1972", "Choudhury_1988", "Su_2001", "constant_kB"),
  kB = NULL, constants = bigleaf.constants())
}
\arguments{
\item{data}{Data.frame or matrix containing all required variables}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{wind}{Wind speed (m s-1)}

\item{ustar}{Friction velocity (m s-1)}

\item{H}{Sensible heat flux (W m-2)}

\item{zr}{Instrument (reference) height (m)}

\item{zh}{Canopy height (m)}

\item{d}{Zero-plane displacement height (m)}

\item{z0m}{Roughness length for momentum (m)}

\item{Dl}{Characteristic leaf dimension (m); only used if \code{Rb_model} is "Choudhury_1988" or "Su_2001".}

\item{N}{Number of leaf sides participating in heat exchange (1 or 2); only used if \code{Rb_model} is "Su_2001".}

\item{fc}{Fractional vegetation cover (-); only used if \code{Rb_model} is "Su_2001". See Details.}

\item{LAI}{One-sided leaf area index (m2 m-2); only used if \code{Rb_model} is "Choudhury_1988" or "Su_2001".}

\item{Cd}{Foliage drag coefficient (-); only used if \code{Rb_model} is "Su_2001".}

\item{hs}{Roughness length of bare soil (m); only used if \code{Rb_model} is "Su_2001".}

\item{stab_correction}{Should stability correction be applied? Default is TRUE}

\item{stab_formulation}{Stability correction function. Either "Dyer_1970" (default) or
"Businger_1971". Ignored if \code{stab_correction} is FALSE.}

\item{Rb_model}{Boundary layer resistance formulation. One of c("Thom_1972","Choudhury_1988","Su_2001).}

\item{kB}{kB-1 value; only used if \code{Rb_model = "constant_kB"}}

\item{constants}{k - von Karman constant (-) \cr
Rbwc - Ratio of the transfer efficiency through the boundary layer for water vapor and CO2 (-) \cr
cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
Kelvin - conversion degree Celsius to Kelvin \cr
g - gravitational acceleration (m s-2) \cr
pressure0 - reference atmospheric pressure at sea level (Pa) \cr
Tair0 - reference air temperature (K) \cr}
}
\value{
a dataframe with the following columns:
        \item{Ga_m}{Aerodynamic conductance for momentum (m s-1)}
        \item{Ga_h}{Aerodynamic conductance for heat and water vapor (m s-1)}
        \item{Ga_CO2}{Aerodynamic conductance for CO2 (m s-1)}
        \item{Ra_m}{Aerodynamic resistance for momentum (s m-1)}
        \item{Ra_h}{Aerodynamic resistance for heat and water vapor (s m-1)}
        \item{Ra_CO2}{Aerodynamic resistance for CO2 (s m-1)}
        \item{Gb}{Canopy boundary layer conductance for heat and water vapor (m s-1)}
        \item{Rb}{Canopy boundary layer resistance for heat and water vapor (s m-1)}
        \item{Rb_CO2}{Canopy boundary layer resistance for CO2 (s m-1)}
        \item{kB}{kB-1 parameter (-)}
        \item{zeta}{Stability parameter 'zeta' (-)}
        \item{psi_h}{Integrated stability correction function (-)}
}
\description{
Bulk aerodynamic conductance, including options for the boundary layer conductance
             formulation and stability correction functions.
}
\details{
Aerodynamic conductance (Ga) is calculated as:

 Ga = 1 / (Ra_m + Rb)
 
 where Ra_m is the aerodynamic resistance for momentum and Rb the canopy boundary
 layer resistance.
 
 Several models can be used to determine the canopy boundary layer resistance ("excess resistance")

 The model used to determine the canopy boundary layer resistance ("excess resistance") is specified by the argument \code{Rb_model}
 "Thom_1972" is an empirical formulation based on the friction velocity (ustar) (Thom 1972):
 
   \deqn{Rb = 0.67ustar^0.67}
   
 The model by Choudhury 1988 ("Choudhury_1988"), calculates Rb
 based on leaf width, LAI and ustar (Note that in the original formulation leaf width
 instead of the characteristic leaf dimension (Dl) is taken):
  
    \deqn{Gb = LAI((2a/\alpha)*sqrt(u(h)/w)*(1-exp(-\alpha/2)))}
    
 The option "Su_2001" calculates Rb based on the physically-based Rb model by Su et tal. 2001,
 a simplification of the model developed by Massmann 1999:
 
    \deqn{kB-1 = (k Cd fc^2) / (4Ct ustar/u(zh)) + kBs-1(1 - fc)^2}
 
 The models calculate the parameter kB-1, which related to Rb:
 
    \deqn{kB-1 = Rb * (k * ustar)}
 
 Both kB and Rb, as well as Gb are given in the output.
 
 Rb (and Gb) for water vapor and heat are assumed to be equal. Rb for CO2 (Rb_CO2) is given as:
 
 \deqn{Rb_CO2 = 1.4 * Rb}
 
 The factor 1.4 arises due the lower molecular diffusivity of CO2 compared to water.
 It is lower than the ratio of the molecular diffusivities (Dw/DCO2 = 1.6), as movement
 across the boundary layer is assumed to be partly by diffusion and partly by turbulent
 mixing (Nobel 2005).
 
 If the roughness parameters z0m and d are not available, they can be estimated using
 \link{\code{roughness.parameters}}.
 
 The argument \code{stab_formulation} determines the stability correction function used 
 to account for the effect of atmospheric stability on Ra_m (Ra_m is lower for unstable
 and higher for stable stratification). Stratification is based on a stability parameter zeta (z-d/L),
 where z = reference height, d the zero-plane displacement height, and L the Monin-Obukhov length, 
 calculated with \code{\link{MoninObukhov.length}}
 Stability correction functions calculate the correction based on the formulations by Businger 1971 
 Dyer 1970, based on zeta.
 If \code{stab_correction} is set to FALSE, the effects of atmospheric stability on Ga are neglected,
 and the Monin-Obukhov length, the stability parameter zeta, as well as the stability correction functions
 are not calculated.
}
\references{

}

