% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{air.density}
\alias{air.density}
\title{Air density}
\usage{
air.density(Tair, pressure, constants)
}
\arguments{
\item{Tair}{air temperature (degC)}

\item{pressure}{air pressure (Pa)}

\item{Mair}{molar mass of air (kg mol-1)}

\item{Rgas}{ideal gas constant ()}
}
\description{
Air density from air temperature and pressure
}

