% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{air.density}
\alias{air.density}
\title{Air density}
\usage{
air.density(Tair, pressure, constants = bigleaf.constants())
}
\arguments{
\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{constants}{Kelvin - conversion degC to Kelvin \cr
Rd - gas constant of dry air (J kg-1 K-1)}
}
\value{
rho air density (kg m-3)
}
\description{
Air density of moist air from air temperature and pressure.
}
\details{
Air density (\eqn{\rho}) is calculated as:
\deqn{\rho = pressure / Rd * Tair}
}
\references{
Foken, T, 2008: Micrometeorology. Springer, Berlin, Germany.
}
