% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf_physiology.r
\name{Arrhenius.temp.response}
\alias{Arrhenius.temp.response}
\title{(Modified) Arrhenius temperature response function}
\usage{
Arrhenius.temp.response(param, Temp, Ha, Hd = NULL, dS = NULL,
  constants = bigleaf.constants())
}
\arguments{
\item{param}{Parameter measured at measurement temperature (umol m-2 s-1)}

\item{Temp}{Measurement temperature (degC)}

\item{Ha}{Activation energy for param (kJ mol-1)}

\item{Hd}{Deactivation energy for param (kJ mol-1)}

\item{dS}{Entropy term for param (J mol-1 K-1)}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
Rgas - universal gas constant (J mol-1 K-1)}
}
\value{
param25 - value of the input parameter at the reference temperature of 25degC (umol m-2 s-1)
}
\description{
Common or modified (i.e. with temperature optimum - if the deactivation
             energy and entropy term are provided) Arrhenius function describing
             the temperature response of biochemical parameters.
}
\details{
\deqn{Vcmax25 = Vcmax / 
                            ( exp(Vcmax_Ha * (Temp - Tref) / (Tref*constants$Rgas*Temp)) *
                            (1 + exp((Tref*dS - Vcmax_Hd) / (Tref * constants$Rgas))) /
                            (1 + exp((Temp*dS - Vcmax_Hd) / (Temp * constants$Rgas)))
                            )
               }
}
\references{
Kattge J., Knorr W., 2007: Temperature acclimation in a biochemical
            model of photosynthesis: a reanalysis of data from 36 species.
            Plant, Cell and Environment 30, 1176-1190.
}
