% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{carbox.rate}
\alias{carbox.rate}
\title{Bulk canopy maximum carboxylation efficiency (Vcmax)}
\usage{
carbox.rate(Temp, GPP, Ci, PPFD, PPFD_sat, O2 = 0.21, Kc25 = 404.9,
  Ko25 = 278.4, Kc_Ha = 79.43, Ko_Ha = 36.38, Vcmax_Ha = 65.33,
  Vcmax_Hd = 200, dS = 640, constants = bigleaf.constants())
}
\arguments{
\item{Temp}{Surface (or air) temperature (degC)}

\item{GPP}{Gross primary productivty (umol m-2 s-1)}

\item{Ci}{Bulk canopy intercellular CO2 concentration (umol mol-1)}

\item{PPFD}{Photosynthetic photon flux density (umol m-2 s-1)}

\item{PPFD_sat}{PPFD threshold, above which the canopy is considered to 
be light saturated.}

\item{Kc25}{Michaelis-Menten constant for CO2 at 25 degC (umol mol-1)}

\item{Ko25}{Michaelis-Menten constant for O2 at 25 degC (mmol mol-1)}

\item{Kc_Ha}{activation energy for Kc (kJ mol-1)}

\item{Ko_Ha}{activation energy for Ko (kJ mol-1)}

\item{Vcmax_Ha}{activation energy for Vcmax (kJ mol-1)}

\item{Vcmax_Hd}{deactivation energy for Vcmax (kJ mol-1)}

\item{dS}{entropy term (J mol-1 K-1)}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
Rgas -}

\item{Ox}{O2 concentration (mol mol-1)}
}
\description{
Bulk canopy maximum carboxylation rate (Vcmax) from bulk intercellular 
             CO2 concentration using the Farquhar et al. 1980 model for C3 photosynthesis.
}
