% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{check.columns}
\alias{check.columns}
\title{checks columns in a data frame}
\usage{
check.columns(data, column_name)
}
\arguments{
\item{data}{a data.frame or a matrix}

\item{column_name}{the column name}
}
\description{

}

