% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.r
\name{check.input}
\alias{check.input}
\title{Check Input for Functions in the bigleaf Package}
\usage{
check.input(data, ...)
}
\arguments{
\item{data}{Input data.frame or matrix}

\item{...}{Input variables. Either a list or individual vectors}
}
\description{
Checks length and type of the provided input variables.
}
\note{
This function can be run for named variables (in which case the return
      value will be named according to the given name), or for placeholder
      variables that are assigned and named according to e.g. entries of a character 
      vector. In the latter case, the input variable has to be named as \code{"var"} or
      \code{"var_qc"}.
}
\keyword{internal}
