% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{check.length}
\alias{check.length}
\title{tests if variables have the same length}
\usage{
check.length(...)
}
\arguments{
\item{...}{variables for which the length has to be compared}
}
\description{
tests if variables have the same length
}
