% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{decoupling}
\alias{decoupling}
\title{Decoupling coefficient}
\usage{
decoupling(data, Tair = "Tair", pressure = "pressure", Ga = "Ga",
  Gs = "Gs", approach = c("JarvisMcNaughton_1986", "Martin_1989"),
  canopy.type = c("amphistomatous", "hypostomatous"), LAI,
  constants = bigleaf.constants())
}
\arguments{
\item{data}{Data.frame or matrix containing all required input variables}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{Ga}{Aerodynamic conductance (m s-1)}

\item{Gs}{Surface conductance (m s-1)}

\item{approach}{Approach used to calculate omega. Either "JarvisMcNaughton_1986" (default)
or "Martin_1989".}

\item{canopy.type}{Either "amphistomatous" (default) or "hypostomatous". See Details.}

\item{LAI}{Leaf area index (m2 m-2), only used if approach = "Martin_1989"}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
eps - ratio of the molecular weight of water vapor to dry air (-) \cr
sigma - Stefan-Boltzmann constant (W m-2 K-4) \cr}
}
\value{
omega The decoupling coefficient omega (-)
}
\description{
The canopy-atmosphere decoupling coefficient 'Omega'.
}
\details{
The decoupling coefficient Omega ranges from 0 to 1 and quantifies the
         linkage of the conditions (foremost humidity and temperature) at the canopy surface
         to the ambient air. Values close to 0 indicate well coupled conditions
         characterized by high physiological (i.e. stomatal) control on transpiration
         and similar conditions at the canopy surface compared to the atmosphere above
         the canopy. Values close to 1 indicate the opposite, i.e. decoupled conditions and 
         a low stomatal control on transpiration (Jarvis & McNaughton 1986). \cr
         The "JarvisMcNaughton_1986" approach (default option) is the original
         formulation for the decoupling coefficient, given by (for an amphistomatous 
         canopy (i.e. stomata located on both sides of leaves)):
         
         \deqn{\Omega = \frac{\epsilon + 1}{\epsilon + 1 + \frac{Ga}{Gc}}}{%
         \Omega = (\epsilon + 1) / ( \epsilon + 1 + Ga/Gc)}
         
         where \eqn{\epsilon = \frac{s}{\gamma}}{\epsilon = s/\gamma} is a dimensionless coefficient
         with s being the slope of the saturation vapor pressure curve (Pa K-1), and \eqn{\gamma} the 
         psychrometric constant (Pa K-1).
         
         This formulation can be modified for for hypostomatous (stomata located 
         on only one side of the leaves) vegetation (e.g. Daudet et al. 1999): 
         
         \deqn{\Omega = \frac{\epsilon + 2}{\epsilon + 2 + 2\frac{Ga}{Gc}}}{%
         \Omega = (\epsilon + 2) / (\epsilon + 2 + 2Ga/Gc)}
         
         The approach "Martin_1989" by Martin 1989 additionally takes radiative coupling
         into account. For amphistomatous vegetation:
         
         \deqn{\Omega = \frac{\epsilon + 1 + \frac{Gr}{Ga}}{\epsilon + (1 + \frac{Ga}{Gs}) (1 + \frac{Gr}{Ga})}}{%
         \Omega = (\epsilon + 1 + Gr/Ga) / (\epsilon + (1 + Ga/Gs) (1 + Gr/Ga))}
}
\references{
Goldberg V., Bernhofer C., 2008: Testing different decoupling coefficients
            with measurements and models of contrasting canopies and soil water conditions.
            Annales Geophysicae 26, 1977-1992.

            Jarvis P.G., McNaughton K.G., 1986: Stomatal control of transpiration:
            scaling up from leaf to region. Advances in Ecological Research 15, 1-49. 
            
            Daudet F. A., Le Roux X., Sinoquet H., and Adam A., 1999:
            Wind speed and leaf boundary conductance variation within
            tree crown, Consequences on leaf-atmosphere coupling and tree
            functions, Agricultural and Forest Meteorology 97, 171-185.
            
            Martin P., 1989: The significance of radiative coupling between
            vegetation and the atmosphere. Agricultural and Forest Meteorology 49, 45-53.
}

