% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{dew.point}
\alias{dew.point}
\title{Dew point temperature}
\usage{
dew.point(Tair, pressure, VPD, constants = bigleaf.constants())
}
\arguments{
\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{constants}{Kelvin - 
Mw - 
Rgas -}
}
\value{
dew point Td (deg C)
}
\description{
calculates the dew point, the temperature to which air must be 
             cooled to become saturated (i.e. e = esat(Td))
}
\details{
Dew point is given by:

        \deqn{Td = T* / (1 - log(e/esat(T*) / A))}
   
         where T* is 0 deg C. A is given by:
         
         \deqn{A = \lambda * Mw / (R*Tair)}
         
         where \eqn{\lambda} is the latent heat of vaporization and Mw the 
         molecular weight of water.
}
\references{
Monteith J.L., Unsworth M.H., 2008: Principles of Environmental Physics.
            3rd edition. Academic Press, London.
}
