% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{dew.point}
\alias{dew.point}
\title{Dew point}
\usage{
dew.point(Tair, pressure, VPD)
}
\arguments{
\item{Tair}{Air temperature (degC)}

\item{pressure}{Atmospheric pressure (kPa)}

\item{VPD}{Vapor pressure deficit (kPa)}
}
\value{
\item{Td -}{dew point temperature (degC)}
}
\description{
calculates the dew point, the temperature to which air must be 
             cooled to become saturated (i.e. e = esat(Td))
}
\details{
Dew point temperature (Td) is defined by:

        \deqn{e = Esat(Td)}
   
         which is solved for Td using \code{\link{nls}}.
}
\examples{
dew.point(25,100,1.5)                

}
\references{
Monteith J.L., Unsworth M.H., 2008: Principles of Environmental Physics.
            3rd edition. Academic Press, London.
}
