% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorological_variables.r
\name{dew.point.solver}
\alias{dew.point.solver}
\title{Solver function for dew point temperature}
\usage{
dew.point.solver(ea, accuracy, Esat.formula, constants = bigleaf.constants())
}
\arguments{
\item{ea}{Air vapor pressure (kPa)}

\item{accuracy}{Accuracy of the result (degC)}

\item{Esat.formula}{Optional: formula to be used for the calculation of esat and the slope of esat.
One of \code{"Sonntag_1990"} (Default), \code{"Alduchov_1996"}, or \code{"Allen_1998"}.
See \code{\link{Esat.slope}}.}

\item{constants}{Pa2kPa - conversion pascal (Pa) to kilopascal (kPa)}
}
\description{
Solver function used in dew.point()
}
\note{
Arguments \code{accuracy} and \code{Esat.formula} are passed to this function by dew.point().
}
\keyword{internal}
