% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{e.to.q}
\alias{e.to.q}
\title{Conversions between humidity measures}
\usage{
e.to.q(e, pressure, constants)
}
\arguments{
\item{e}{vapor pressure (Pa)}

\item{p}{air pressure (Pa)}

\item{q}{specific humidity (kg kg-1)}

\item{VPD}{vapore pressure deficit (kPa)}
}
\description{
Conversion between VPD, specific humidity, and relative humidity
}

