% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{e.to.q}
\alias{e.to.VPD}
\alias{e.to.q}
\alias{q.to.e}
\title{Conversions between humidity measures}
\usage{
e.to.q(e, pressure, constants = bigleaf.constants())

q.to.e(q, pressure)

e.to.VPD(e, pressure)
}
\arguments{
\item{e}{vapor pressure (kPa)}

\item{q}{specific humidity (kg kg-1)}

\item{p}{air pressure (kPa)}

\item{VPD}{vapor pressure deficit (kPa)}
}
\description{
Conversion between actual humidity, specific humidity, relative humidity, and VPD
}
\references{
Foken_2008
}

