% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{energy.closure}
\alias{energy.closure}
\title{Energy balance closure}
\usage{
energy.closure(data, Rn = "Rn", G = NULL, S = NULL, LE = "LE",
  H = "H", instantaneous = F, missing.G.as.NA = F, missing.S.as.NA = F)
}
\arguments{
\item{data}{Data.frame or matrix containing all required variables}

\item{Rn}{Net radiation (W m-2)}

\item{G}{Ground heat flux (W m-2); optional}

\item{S}{Sum of all storage fluxes (W m-2); optional}

\item{LE}{Latent heat flux (W m-2)}

\item{H}{Sensible heat flux (W m-2)}
}
\value{
a named vector containing:
        \item{n}{number of complete (all energy balance terms available) observations}
        \item{intercept}{intercept of the OLS regression}
        \item{slope}{slope of the OLS regression}
        \item{r_squared}{r^2 of the OLS regression}
        \item{EBR}{energy balance ratio}
        
        if \code{instantaneous} is TRUE, only \code{EBR} is returned
}
\description{
Calculates the degree of the energy balance non-closure for the entire timespan
             based on the ratio of two sums (energy balance ratio), and ordinary least squares (OLS).
}
\details{
The energy balance ratio (EBR) is calculated as:
         
         \deqn{EBR = sum(LE + H)/sum(Rn - G - S)}
         
         the sum is taken for all timesteps with complete observations (i.e. where
         all energy balance terms are available).
}
\references{
Wilson K., et al. 2002: Energy balance closure at FLUXNET sites.
            Agricultural and Forest Meteorology 113, 223-243.
}
