% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.r
\name{energy.use.efficiency}
\alias{energy.use.efficiency}
\title{Energy use efficiency (EUE)}
\usage{
energy.use.efficiency(GPP, alpha = 0.422, Rn)
}
\arguments{
\item{GPP}{Gross ecosystem productivity (umol CO2 m-2 s-1)}

\item{alpha}{Energy taken up/released by photosynthesis/respiration (J umol-1)}

\item{Rn}{Net radiation (W m-2)}
}
\value{
\item{EUE -}{Energy use efficiency (-)}
}
\description{
Fraction of net radiation fixed by biochemical energy.
}
\examples{
energy.use.efficiency(GPP=20,Rn=500)

}
