% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Esat}
\alias{Esat}
\title{Saturation vapor pressure and slope of the curve}
\usage{
Esat(Tair, formula = c("Alduchov_1996", "Sonntag_1990"))
}
\arguments{
\item{Tair}{air temperature (deg C)}

\item{formula}{formula to be used. Either "Alduchov_1996" or "Sonntag_1990"}
}
\value{
A dataframe with the following columns:
\itemize{
  \item{Esat} { - Saturation vapor pressure (Pa)}
  \item{delta}{ - Slope of the saturation vapor pressure curve (Pa K-1)}
}
}
\description{
Calculates Saturation vapor pressure (Esat) over water and the
             corresponding slope of the saturation vapor pressure curve.
}
\details{
Esat (Pa) is calculated based on the Magnus equation:

 \deqn{Esat = a * exp((b * Tair) / (c + Tair))}
 
 where the coefficients a, b, c differ with the formula.
 The slope of the Esat curve is calculated as the first derivative of the function.
}
\references{
Alduchov, O. A. & Eskridge, R. E., 1996: Improved Magnus form approximation of 
            saturation vapor pressure. Journal of Applied Meteorology, 1996, 35, 601-609

            Sonntag_1990 Important new values of the physical constants of 1986, vapor 
            pressure formulations based on the ITS-90, and psychrometric formulae 
            (--> as used by WMO 2008!)
}

