% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{estimate.g1}
\alias{estimate.g1}
\title{Estimation of g1}
\usage{
estimate.g1(data, GPP = "GPP", Gs = "Gs", VPD = "VPD", Ca = "Ca",
  fitg0 = FALSE, nmin = 40, g0 = 0)
}
\arguments{
\item{data}{a data.frame or matrix containing all required columns}

\item{GPP}{gross primary productivity (umol m-2 s-1)}

\item{Gs}{surface conductance (mol m-2 s-1)}

\item{VPD}{vapor pressure deficit (kPa)}

\item{Ca}{atmospheric CO2 concentration (ppm)}

\item{fitg0}{logical, if TRUE g0 and g1 are fitted simultaneously}

\item{nmin}{minimum number of data required to perform the fit}

\item{g0}{ignored if fitg0 is FALSE, minimum stomatal conductance (mol m-2 s-1)}
}
\value{
an nls model object
}
\description{
Estimation of the WUE metric "g1" from non-linear regression
}
\references{
Medlyn 2011
            Knauer 2017
}

