% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{ET.pot}
\alias{ET.pot}
\title{Potential evapotranspiration from the Priestley-Taylor equation}
\usage{
\method{ET}{pot}(data, Tair = "Tair", pressure = "pressure", Rn = "Rn",
  G = NULL, S = NULL, alpha = 1.26, missing.G.as.NA = F,
  missing.S.as.NA = F, constants = bigleaf.constants())
}
\arguments{
\item{data}{data.frame or matrix containing all required columns}

\item{Tair}{air temperature (deg C)}

\item{pressure}{air pressure (kPa)}

\item{Rn}{net radiation (W m-2)}

\item{G}{ground heat flux (W m-2)}

\item{S}{sum of all storage fluxes (W m-2)}

\item{alpha}{Priestley-Taylor coefficient (-)}

\item{constants}{cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
eps - ratio of the molecular weight of water vapor to dry air (-)}
}
\value{
ET_pot potential evapotranspiration (kg m-2 s-1)
}
\description{
Potential evapotranspiration from the Priestley-Taylor equation
}
\details{


potential Evapotranspiration is calculated according to Priestley & Taylor, 1972:

         \deqn{(\alpha * \delta * (Rn - G - S)) / (\delta + \gamma)}
}
\references{
Priestley C.H.B., Taylor R.J., 1972: On the assessment of surface heat flux
            and evaporation using large-scale parameters. Monthly Weather Review 100, 81-92.
}

