% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{ET.pot}
\alias{ET.pot}
\title{Potential evapotranspiration}
\usage{
ET.pot(data, Tair = "Tair", pressure = "pressure", Rn = "Rn", G = NULL,
  S = NULL, alpha = 1.26, missing.G.as.NA = F, missing.S.as.NA = F,
  constants = bigleaf.constants())
}
\arguments{
\item{data}{Data.frame or matrix containing all required variables; optional}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Atmospheric pressure (kPa)}

\item{Rn}{Net radiation (W m-2)}

\item{G}{Ground heat flux (W m-2); optional}

\item{S}{Sum of all storage fluxes (W m-2); optional}

\item{alpha}{Priestley-Taylor coefficient (-)}

\item{missing.G.as.NA}{if TRUE, missing G are treated as NA,otherwise set to 0.}

\item{missing.S.as.NA}{if TRUE, missing S are treated as NA,otherwise set to 0.}

\item{constants}{cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
eps - ratio of the molecular weight of water vapor to dry air (-)}
}
\value{
a data.frame with the following columns:
        \item{ET_pot}{potential evapotranspiration (kg m-2 s-1)}
        \item{LE_pot}{potential latent heat flux (W m-2)}
}
\description{
Potential evapotranspiration ET_pot according to Priestley & Taylor 1972.
}
\details{
Potential evapotranspiration is calculated according to Priestley & Taylor, 1972:

         \deqn{LE_pot = (\alpha * \Delta * (Rn - G - S)) / (\Delta + \gamma)}
}
\note{
If the first argument 'data' is provided (either a matrix or a data.frame),
      the following variables can be provided as character (in which case they are interpreted as
      the column name of 'data') or as numeric vectors, in which case they are taken
      directly for the calculations. If 'data' is not provided, all input variables have to be
      numeric vectors.
}
\examples{
# Calculate potential ET from a surface that receives Rn of 400 Wm-2
ET.pot(Tair=30,pressure=100,Rn=400,alpha=1.26)    

}
\references{
Priestley C.H.B., Taylor R.J., 1972: On the assessment of surface heat flux
            and evaporation using large-scale parameters. Monthly Weather Review 100, 81-92.
}
