% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{filter.data}
\alias{filter.data}
\title{Filter data}
\usage{
filter.data(data, precip = "precip", PPFD = "PPFD", Tair = "Tair",
  ustar = "ustar", VPD = "VPD", GPP = "GPP_nt", Ca = "Ca",
  day = "day", month = "month", year = "year", quality_ext = "_qc",
  good_quality = c(0, 1), excl_vars_qc = c("PPFD", "Ca"), tprecip = 0.01,
  precip_hours = 24, trad = 200, ttemp = 5, tustar = 0.2, tGPP = 0.5,
  trH = 0.95, NA.as.precip = F)
}
\arguments{
\item{data}{data.frame or matrix containing all required input variables in 
half-hourly or hourly resolution. Including year, month, day information}

\item{precip}{precipitation (mm)}

\item{PPFD}{photosynthetically photon flux density (umol m-2 s-1)}

\item{Tair}{air temperature (deg C)}

\item{ustar}{friction velocity (m s-1)}

\item{VPD}{vapor pressure deficit (kPa)}

\item{GPP}{gross primary productivity (umol m-2 s-1)}

\item{Ca}{atmospheric CO2 concentration (umol mol-1)}

\item{day}{day of year}

\item{month}{month}

\item{year}{year}

\item{quality_ext}{}

\item{good_quality}{}

\item{excl_vars_qc}{}

\item{tprecip}{precipitation threshold used to demark a precipitation event (mm)}

\item{precip_hours}{number of hours removed following a precipitation event (h)}

\item{trad}{radiation threshold (umol m-2 s-1)}

\item{ttemp}{temperature threshold (deg C)}

\item{tustar}{friction velocity threshold (m s-1)}

\item{tGPP}{GPP threshold (???)}

\item{trH}{relative humidity threshold (-)}

\item{NA.as.precip}{if TRUE missing precipitation data are treated as precipitation events}

\item{rH}{relative humidity (-), only one of VPD and rH required}
}
\value{
filtered_data the same data frame as provided as argument to the function
                      but with filtered timeperiods set to NA in all columns except
                      the ones used for filtering. The data.frame contains one additional
                      column "valid", which indicates whether the timestep is valid 
                      according to the filter criteria (1) or invalid (0).
}
\description{
Filters timeseries in which Gs describes physiologically meaningful
             quantity
}
\details{
growing season:
         growing season is filtered based on daily GPP (aggregated from halfhourly values).
}
\references{
Knauer 2017
}

