% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{filter.data}
\alias{filter.data}
\title{Basic data filtering}
\usage{
filter.data(data, precip = "precip", PPFD = "PPFD", Tair = "Tair",
  ustar = "ustar", VPD = "VPD", GPP = "GPP_nt", doy = "doy",
  year = "year", quality.control = TRUE, filter.growseas = FALSE,
  quality.ext = "_qc", vars.qc = c("precip", "Tair", "VPD", "GPP_nt", "H",
  "LE"), good.quality = c(0, 1), missing.qc.as.bad = TRUE, tprecip = 0.01,
  precip.hours = 24, NA.as.precip = FALSE, trad = 200, ttemp = 5,
  tustar = 0.2, trH = 0.95, tGPP = 0.5, ws = 15, min.int = 5)
}
\arguments{
\item{data}{Data.frame or matrix containing all required input variables in 
half-hourly or hourly resolution. Including year, month, day information}

\item{precip}{Precipitation (mm)}

\item{PPFD}{Photosynthetically photon flux density (umol m-2 s-1)}

\item{Tair}{Air temperature (deg C)}

\item{ustar}{Friction velocity (m s-1)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{GPP}{Gross primary productivity (umol m-2 s-1); Ignored if \code{filter.growseas} is FALSE.}

\item{doy}{Day of year}

\item{year}{Year}

\item{quality.control}{Should quality control be applied? Defaults to TRUE}

\item{filter.growseas}{Should data be filtered for growing season? Defaults to FALSE}

\item{quality.ext}{The extension to the variables' names that marks them as 
quality control variables. Ignored if \code{quality.control} is FALSE.}

\item{vars.qc}{Character vector indicating the variables for which quality filter should 
be applied. Ignored if \code{quality.control} is FALSE.}

\item{good.quality}{Which values indicate good quality in the quality control (qc) variables?
Ignored if \code{quality.control} is FALSE.}

\item{missing.qc.as.bad}{If quality control variable is NA, should the corresponding data point be
treated as bad quality? Defaults to TRUE. Ignored if \code{quality.control} is FALSE.}

\item{tprecip}{Precipitation threshold used to demark a precipitation event (mm)}

\item{precip.hours}{Number of hours removed following a precipitation event (h)}

\item{NA.as.precip}{If TRUE missing precipitation data are treated as precipitation events}

\item{trad}{Radiation threshold (umol m-2 s-1)}

\item{ttemp}{Temperature threshold (deg C)}

\item{tustar}{Friction velocity threshold (m s-1)}

\item{trH}{Relative humidity threshold (-). Note that relative humidity
is calculated from VPD internally}

\item{tGPP}{GPP threshold (fraction of 95% quantile of the GPP time series).
Must be between 0 and 1. Ignored if \code{filter.growseas} is FALSE.}

\item{ws}{Window size used for GPP time series smoothing. 
Ignored if \code{filter.growseas} is FALSE.}

\item{min.int}{Minimum time interval in days for a given state of growing season.
Ignored if \code{filter.growseas} is FALSE.}
}
\value{
The same data frame as provided as first argument to the function
        with an additional column "valid", which indicates whether the timestep fulfills 
        the filtering criteria (1) or not (0).
}
\description{
Filters timeseries of EC data for high-quality values and specified
             meteorological conditions.
}
\details{
This routine consists of two parts:

         1) Quality control: All variables included in \code{vars.qc} are filtered for 
            good quality data. For these variables a corresponding quality variable with 
            the same name as the variable plus the extension as specified in \code{quality.ext}
            must be provided. For timesteps where the value of the quality indicator is not included
            in the argument \code{good.quality}, i.e. the quality is not considered as 'good', 
            its value is set to NA.
            
         2) Meteorological filtering. Under certain conditions (e.g. low ustar), the assumptions
            of the EC method are not fulfilled. Further, some data analysis require certain meteorological
            conditions, such as precipitation free periods, or active vegetation (growing season, daytime).
            The filter applied in this second step serves to exclude time periods that do not fulfill the criteria
            specified in the arguments. More specifically, timeperiods where one of the variables is lower than the 
            specified threshold are set to NA for all variables. E.g. when radiation is below 'trad', all variables 
            are set to NA.
            Set a threshold to 0 (or the minimum of the variable) if the dataset should not be filtered for a specific variable.
}
\note{
This function is not exhaustive, and one may wish to exclude further situations 
      (e.g. negative LE fluxes, counter-gradient fluxes, etc.), depending on the purpose.
}
\examples{
# Example of data filtering; data are for a month within the growing season,
# hence growing season is not filtered.
DE_Tha_Jun_2014 <- filter.data(DE_Tha_Jun_2014,quality.control=TRUE,filter.growseas=FALSE,
                               quality.ext="_qc",vars.qc=c("precip","Tair","VPD","H","LE"),
                               tprecip=0.01,precip.hours=24,NA.as.precip=FALSE,trad=200,
                               ttemp=5,tustar=0.2,tGPP=0,ws=15,min.int=5,trH=0.95)
# note the additional column 'valid' in DE_Tha_Jun_2014.                                 

}
