% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.r
\name{filter.data}
\alias{filter.data}
\title{Basic data filtering}
\usage{
filter.data(data, quality.control = TRUE, vars.qc = NULL,
  filter.growseas = FALSE, filter.precip = FALSE, filter.vars = NULL,
  filter.vals.min, filter.vals.max, NA.as.invalid = TRUE,
  quality.ext = "_qc", good.quality = c(0, 1), missing.qc.as.bad = TRUE,
  GPP = "GPP", doy = "doy", year = "year", tGPP = 0.5, ws = 15,
  min.int = 5, precip = "precip", tprecip = 0.01, precip.hours = 24,
  records.per.hour = 2, filtered.data.to.NA = TRUE)
}
\arguments{
\item{data}{Data.frame or matrix containing all required input variables in 
half-hourly or hourly resolution. Including year, month, day information}

\item{quality.control}{Should quality control be applied? Defaults to TRUE.}

\item{vars.qc}{Character vector indicating the variables for which quality filter should 
be applied. Ignored if \code{quality.control} is FALSE.}

\item{filter.growseas}{Should data be filtered for growing season? Defaults to FALSE.}

\item{filter.precip}{Should precipitation filtering be applied? Defaults to FALSE.}

\item{filter.vars}{Additional variables to be filtered. Vector of type character.}

\item{filter.vals.min}{Minimum values of the variables to be filtered. Numeric vector of 
the same length than \code{filter.vars}. Set to NA to be ignored.}

\item{filter.vals.max}{Maximum values of the variables to be filtered. Numeric vector of 
the same length than \code{filter.vars}. Set to NA to be ignored.}

\item{NA.as.invalid}{If TRUE (the default) missing data are filtered out (applies to all variables).}

\item{quality.ext}{The extension to the variables' names that marks them as 
quality control variables. Ignored if \code{quality.control} is FALSE.}

\item{good.quality}{Which values indicate good quality (i.e. not to be filtered) 
in the quality control (qc) variables? Ignored if \code{quality.control} is FALSE.}

\item{missing.qc.as.bad}{If quality control variable is NA, should the corresponding data point be
treated as bad quality? Defaults to TRUE. Ignored if \code{quality.control} is FALSE.}

\item{GPP}{Gross primary productivity (umol m-2 s-1); Ignored if \code{filter.growseas} is FALSE.}

\item{doy}{Day of year; Ignored if \code{filter.growseas} is FALSE.}

\item{year}{Year; Ignored if \code{filter.growseas} is FALSE.}

\item{tGPP}{GPP threshold (fraction of 95% quantile of the GPP time series).
Must be between 0 and 1. Ignored if \code{filter.growseas} is FALSE.}

\item{ws}{Window size used for GPP time series smoothing. 
Ignored if \code{filter.growseas} is FALSE.}

\item{min.int}{Minimum time interval in days for a given state of growing season.
Ignored if \code{filter.growseas} is FALSE.}

\item{precip}{Precipitation (mm time-1)}

\item{tprecip}{Precipitation threshold used to demark a precipitation event (mm). 
Ignored if \code{filter.precip} is FALSE}

\item{precip.hours}{Number of hours removed following a precipitation event (h).
Ignored if \code{filter.precip} is FALSE}

\item{records.per.hour}{Number of observations per hour. I.e. 2 for half-hourly data.}

\item{filtered.data.to.NA}{Logical. If TRUE (the default), all variables in the input
data.frame/matrix are set to NA for the timestep where ANY of the
\code{filter.vars} were beyond their acceptable range (as
determined by \code{filter.vals.min} and \code{filter.vals.max}).
If FALSE, values are not filtered, and an additional column 'valid'
is added to the data.frame/matrix, indicating if any value of a row
did (1) or did not fulfill the filter criteria (0).}
}
\value{
If \code{filtered.data.to.NA==TRUE} (default), the input data.frame/matrix with 
        observations which did not fulfill the filter criteria set to NA. 
        If \code{filtered.data.to.NA==FALSE}, the input data.frame/matrix with an additional 
        column "valid", which indicates whether all the data of a timestep fulfill the 
        filtering criteria (1) or not (0).
}
\description{
Filters timeseries of EC data for high-quality values and specified
             meteorological conditions.
}
\details{
This routine consists of two parts:

         1) Quality control: All variables included in \code{vars.qc} are filtered for 
            good quality data. For these variables a corresponding quality variable with 
            the same name as the variable plus the extension as specified in \code{quality.ext}
            must be provided. For timesteps where the value of the quality indicator is not included
            in the argument \code{good.quality}, i.e. the quality is not considered as 'good', 
            its value is set to NA.
            
         2) Meteorological filtering. Under certain conditions (e.g. low ustar), the assumptions
            of the EC method are not fulfilled. Further, some data analysis require certain meteorological
            conditions, such as periods without rainfall, or active vegetation (growing season, daytime).
            The filter applied in this second step serves to exclude time periods that do not fulfill the criteria
            specified in the arguments. More specifically, timeperiods where one of the variables is higher
            or lower than the specified thresholds (\code{filter.vals.min} and \code{filter.vals.max})
            are set to NA for all variables. If a threshold is set to NA, it will be ignored.
}
\note{
Variables considered of bad quality (as specified by the corresponding quality control variables)      
      will be set to NA by this routine. Data that do not fulfill the filtering critera are set to
      NA if \code{filtered.data.to.NA==TRUE}. Note that with this option *all* variables of the same
      time step are set to NA. Alternatively, if \code{filtered.data.to.NA==FALSE} data are not set to NA,
      and a new column "valid" is added to the data.frame/matrix, indicating if any value of a row
      did (1) or did not fulfill the filter criteria (0).
}
\examples{
# Example of data filtering; data are for a month within the growing season,
# hence growing season is not filtered.
DE_Tha_Jun_2014_2 <- filter.data(DE_Tha_Jun_2014,quality.control=FALSE,
                                 vars.qc=c("Tair","precip","H","LE"),
                                 filter.growseas=FALSE,filter.precip=TRUE,
                                 filter.vars=c("Tair","PPFD","ustar"),
                                 filter.vals.min=c(5,200,0.2),
                                 filter.vals.max=c(NA,NA,NA),NA.as.invalid=TRUE,
                                 quality.ext="_qc",good.quality=c(0,1),
                                 missing.qc.as.bad=TRUE,GPP="GPP_nt",doy="doy",
                                 year="year",tGPP=0.5,ws=15,min.int=5,precip="precip",
                                 tprecip=0.1,precip.hours=24,records.per.hour=2)

 # note the additional column 'valid' in DE_Tha_Jun_2014_2.
 # To remove timesteps marked as filtered out (i.e. 0 values in column 'valid'):
 DE_Tha_Jun_2014_2[DE_Tha_Jun_2014_2["valid"] == 0,] <- NA
  
  
}
