% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{filter.data}
\alias{filter.data}
\title{Filter data}
\usage{
filter.data(data, precip = "precip", PPFD = "PPFD", Tair = "Tair",
  ustar = "ustar", VPD = "VPD", GPP = "GPP_nt", doy = "doy",
  year = "year", quality.control = TRUE, quality.ext = "_qc",
  vars.qc = c("precip", "Tair", "VPD", "GPP_nt", "H", "LE"),
  good.quality = c(0, 1), missing.qc.as.bad = TRUE, tprecip = 0.01,
  precip.hours = 24, NA.as.precip = F, trad = 200, ttemp = 5,
  tustar = 0.2, tGPP = 0.5, ws = 15, min.int = 5, trH = 0.95)
}
\arguments{
\item{data}{data.frame or matrix containing all required input variables in 
half-hourly or hourly resolution. Including year, month, day information}

\item{precip}{precipitation (mm)}

\item{PPFD}{photosynthetically photon flux density (umol m-2 s-1)}

\item{Tair}{air temperature (deg C)}

\item{ustar}{friction velocity (m s-1)}

\item{VPD}{vapor pressure deficit (kPa)}

\item{GPP}{gross primary productivity (umol m-2 s-1)}

\item{doy}{day of year}

\item{year}{year}

\item{quality.control}{should quality control be applied?}

\item{quality.ext}{the extension to the variables' names that marks them as 
quality control variables. Ignored if \code{quality.control} is FALSE.}

\item{vars.qc}{character vector indicating the variables for which quality filter should 
be applied. Leave empty if no quality filter should be applied.}

\item{good.quality}{which values indicate good quality in the quality control (qc) variables?}

\item{missing.qc.as.bad}{If quality control variable is NA, should the corresponding data point be
treated as bad quality? Defaults to TRUE}

\item{tprecip}{precipitation threshold used to demark a precipitation event (mm)}

\item{precip.hours}{number of hours removed following a precipitation event (h)}

\item{NA.as.precip}{if TRUE missing precipitation data are treated as precipitation events}

\item{trad}{radiation threshold (umol m-2 s-1)}

\item{ttemp}{Temperature threshold (deg C)}

\item{tustar}{Friction velocity threshold (m s-1)}

\item{tGPP}{GPP threshold (fraction of 95% quantile of the GPP time series).
Must be between 0 and 1.}

\item{ws}{Window size used for GPP time series smoothing}

\item{min.int}{Minimum time interval in days for a given state of growing season}

\item{trH}{Relative humidity threshold (-). Note that relative humidity
is calculated from VPD internally}
}
\value{
The same data frame as provided as first argument to the function
        with an additional column "valid", which indicates whether the timestep fulfills 
        the filtering criteria (1) or not (0).
}
\description{
Filters timeseries of EC data for high-quality values and specified
             meteorological conditions.
}
\details{
This routine consists of two parts:

         1) Quality control: All variables included in \code{vars.qc} are filtered for 
            good quality data. For these variables a corresponding quality variable with 
            the same name as the variable plus the extension as specified in \code{quality.ext}
            must be provided. For timesteps where the value of the quality indicator is not included
            in the argument \code{good.quality}, i.e. the quality is not considered as 'good', 
            its value is set to NA.
            
         2) Meteorological filtering. Under certain conditions (e.g. low ustar), the assumptions
            of the EC method are not fulfilled. Further, some data analysis require certain meteorological
            conditions, such as precipitation free periods, or active vegetation (growing season, daytime).
            The filter applied in this second step serves to exclude time periods that do not fulfill the criteria
            specified in the arguments. More specifically, timeperiods where one of the variables is lower than the 
            specified threshold are set to NA for all variables. E.g. when radiation is below 'trad', all variables 
            are set to NA.
            Set a threshold to 0 (or the minimum of the variable) if the dataset should not be filtered for a specific variable.
}
\examples{
# Example of data filtering; data are for a month within the growing season,
# hence growing season is not filtered (tGPP=0)
DE_Tha_2010 <- filter.data(DE_Tha_Jun_2014,quality.control=TRUE,quality.ext="_qc",
                           vars.qc=c("precip","Tair","VPD","GPP_nt","H","LE"),
                           tprecip=0.01,precip.hours=24,NA.as.precip=F,trad=200,
                           ttemp=5,tustar=0.2,tGPP=0,ws=15,min.int=5,trH=0.95) 

}
