% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Ga}
\alias{Ga}
\title{Aerodynamic conductance}
\usage{
Ga(Tair, presssure, wind, ustar, constants, H, zr, zh, disp, z0m, Dl, N, LAI,
  fc = NULL, Cd = 0.2, hs = 0.01, stab_correction = c("Dyer_1970",
  "Businger_1971", "none"), Rb_model = c("Thom_1972", "McNaughton_1995",
  "Su_2001"))
}
\arguments{
\item{Tair}{air temperature (degC)}

\item{wind}{wind speed (m s-1)}

\item{ustar}{friction velocity (m s-1)}

\item{H}{sensible heat flux (W m-2)}

\item{zr}{measurement (=reference) height (m)}

\item{zh}{canopy height (m)}

\item{disp}{zero-plane displacement height (m)}

\item{z0m}{roughness length for momentum (m)}

\item{Dl}{characteristic leaf dimension (m)}

\item{N}{number of leaf sides participating in heat exchange (1 or 2)}

\item{LAI}{one-sided leaf area index (m2 m-2)}

\item{fc}{fractional vegetation cover (-)}

\item{Cd}{foliage drag coefficient (-)}

\item{hs}{roughness length of bare soil (m)}

\item{stab_correction}{stability correction function}

\item{Rb_model}{boundary layer resistance formulation}

\item{pressure}{air pressure (Pa)}
}
\value{
a dataframe with the following columns:
}
\description{
Aerodynamic conductance, including options for Boundary layer conductance
             and stability correction functions.
}
\details{
Aerodynamic conductance (Ga) is calculated as:

 Ga = 1/Ra_m + 1/Rb
 
 where Ra_m is the aerodynamic resistance for momentum and Rb the canopy boundary
 layer resistance.
 
 The argument stab_correction determines the stability correction function used 
 to account for the effect of atmospheric stability on Ra_m (Ra_m is lower for unstable
 and higher for stable stratification). Stratification is based on a stability parameter zeta (z-d/L),
 where z = reference height, d the zero-plane displacement height, and L the Monin-Obukhov length.
 Stability correction functions calculate the correction based on the formulations by Businger 1971 
 Dyer 1970, based on zeta.
 If "none", atmospheric stability is neglected, and Monin-Obukhov length, and the 
 stability parameter zeta are not calculated.

 Rb_model determines the boundary layer resistance ("excess resistance") model used. Thom_1972 is an 
 empirical formulation based on the friction velocity (ustar):
 
   formula
   
 The model by Mcnaughton and van den Hurk 1995 (McNaughton_1995), calculates Rb
 based on leaf width, LAI and ustar (Note that the original formulation leaf width
 instead of the characteristic leaf dimension (Dl) is taken):
  
    formula
    
 The Su_2001 option calculates Rb based on the physically-based Rb model by Su et tal. 2001,
 a simplification of the model developed by Massmann 1999:
 
    formula
 
 The models calculate the ... parameter kB-1, which related to Rb:
 
    formula
 
 Both kB and Rb are given in the output.
}

