% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Gb.Choudhury}
\alias{Gb.Choudhury}
\title{Boundary layer conductance according to Choudhury & Monteith 1988}
\usage{
Gb.Choudhury(data, Tair = "Tair", pressure = "pressure", wind = "wind",
  ustar = "ustar", H = "H", leafwidth, LAI, zh, zr, d, z0m,
  formulation = c("Dyer_1970", "Businger_1971"),
  constants = bigleaf.constants())
}
\arguments{
\item{wind}{wind speed at sensor height (m s-1)}

\item{ustar}{friction velocity (m s-1)}

\item{leafwidth}{leaf width (m)}

\item{LAI}{one-sided leaf area index}

\item{zh}{canopy height (m)}

\item{zr}{sensor (instrument) height (m)}

\item{constants}{k - von-Karman constant (-)}
}
\value{
a data frame with the following columns:
    \item{Rb}{Boundary layer resistance (s m-1)}
    \item{Gb}{Boundary layer conductance (m s-1)}
    \item{kB}{kB-1 parameter (-)}
}
\description{
Boundary layer conductance according to Choudhury & Monteith 1988
}
\details{
Boundary layer conductance according to Choudhury & Monteith 1988 is
         given by:
         
         \deqn{Gb = LAI((2a/\alpha)*sqrt(u(h)/w)*(1-exp(-\alpha/2)))}
         
         where u(zh) is the wind speed at the canopy surface, approximated from
         measured wind speed at sensor height zr and a wind extinction coefficient \eqn{\alpha}:
         
         \deqn{u(zh) = u(zr) / (exp(\alpha(zr/zh -1)))}.
         
         \eqn{\alpha} is modeled as an empirical relation to LAI (McNaughton & van den Hurk 1995):
         
         \deqn{\alpha = 4.39 - 3.97*exp(-0.258*LAI)}
}
\references{
Choudhury, B. J., Monteith J.L., 1988: A four-layer model for the heat
            budget of homogeneous land surfaces. Q. J. R. Meteorol. Soc. 114, 373-398.
            
            McNaughton, K. G., Van den Hurk, B.J.J.M., 1995: A 'Lagrangian' revision of
            the resistors in the two-layer model for calculating the energy budget of a
            plant canopy. Boundary-Layer Meteorology 74, 261-288.
}
\seealso{
\code{\link{Gb.Thom}}, \code{\link{Gb.Su}}
}

