% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Gb.Su}
\alias{Gb.Su}
\title{Boundary layer conductance according to Su et al. 2001}
\usage{
Gb.Su(data, Tair = "Tair", pressure = "pressure", ustar = "ustar",
  wind = "wind", H = "H", zh, Dl, N, fc = NULL, LAI, Cd = 0.2,
  hs = 0.01, Pr = 0.71, constants = bigleaf.constants())
}
\arguments{
\item{data}{a matrix or data.frame containing all required variables}

\item{Tair}{air temperature (degC)}

\item{ustar}{friction velocity (m s-1)}

\item{wind}{wind speed (m s-1)}

\item{H}{sensible heat flux (W m-2)}

\item{zh}{canopy height (m)}

\item{Dl}{leaf dimension (m)}

\item{N}{number of leaf sides participating in heat exchange (1 or 2)}

\item{fc}{fractional vegetation cover [0-1] (if not provided, calculated from LAI)}

\item{LAI}{one-sided leaf area index (-)}

\item{Cd}{foliage drag coefficient (-)}

\item{hs}{roughness height of the soil (m)}

\item{Pr}{Prandtl number (-), defaults to 0.71}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
pressure0 - reference atmospheric pressure at sea level (Pa) \cr
Tair0 - reference air temperature (K)}

\item{presssure}{atmospheric pressure (Pa)}
}
\value{
a data.frame with the following columns:
    \item{Rb}{Boundary layer resistance (s m-1)}
    \item{Gb}{Boundary layer conductance (m s-1)}
    \item{kB}{kB-1 parameter (-)}
}
\description{
A physically based formulation for the canopy boundary layer conductance.
}
\references{
Su, Z., Schmugge, T., Kustas, W. & Massman, W., 2001: An evaluation of
            two models for estimation of the roughness height for heat transfer between
            the land surface and the atmosphere. Journal of Applied Meteorology 40, 1933-1951.

            Massman, W., 1999: A model study of kB H- 1 for vegetated surfaces using
           'localized near-field' Lagrangian theory. Journal of Hydrology 223, 27-43.
}
\seealso{
\code{\link{Gb.Thom}}, \code{\link{Gb.Choudhury}}
}

