% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Gb.Su}
\alias{Gb.Su}
\title{Boundary layer conductance according to Su et al. 2001}
\usage{
Gb.Su(data, Tair = "Tair", pressure = "pressure", ustar = "ustar",
  wind = "wind", H = "H", zh, zr, d, z0m, Dl, N, fc = NULL, LAI,
  Cd = 0.2, hs = 0.01, stab_correction = TRUE,
  stab_formulation = c("Dyer_1970", "Businger_1971"),
  constants = bigleaf.constants())
}
\arguments{
\item{data}{Data.frame or matrix containing all required variables}

\item{Tair}{air temperature (degC)}

\item{ustar}{friction velocity (m s-1)}

\item{wind}{wind speed (m s-1)}

\item{H}{sensible heat flux (W m-2)}

\item{zh}{canopy height (m)}

\item{zr}{reference height (m)}

\item{d}{zero-plane displacement height (m)}

\item{z0m}{roughness length for momentum (m)}

\item{Dl}{leaf characteristic dimension (m)}

\item{N}{number of leaf sides participating in heat exchange (1 or 2)}

\item{fc}{fractional vegetation cover [0-1] (if not provided, calculated from LAI)}

\item{LAI}{one-sided leaf area index (-)}

\item{Cd}{foliage drag coefficient (-)}

\item{hs}{roughness height of the soil (m)}

\item{stab_correction}{should stability correction be applied? Defaults to TRUE}

\item{stab_formulation}{stability correction function used (If stab_correction is TRUE).
Either "Dyer_1970" or "Businger_1971".}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
pressure0 - reference atmospheric pressure at sea level (Pa) \cr
Tair0 - reference air temperature (K) \cr
Rbwc - Ratio of the transfer efficiency through the boundary layer for water vapor and CO2 (-)}

\item{presssure}{atmospheric pressure (Pa)}
}
\value{
a data.frame with the following columns:
    \item{Rb}{Boundary layer resistance for heat and water (s m-1)}
    \item{Rb_CO2}{Boundary layer resistance for CO2 (s m-1)}
    \item{Gb}{Boundary layer conductance (m s-1)}
    \item{kB}{kB-1 parameter (-)}
}
\description{
A physically based formulation for the canopy boundary layer conductance.
}
\details{
The formulation is based on the kB-1 model developed by Massman 1999. 
         Su et al. 2001 derived the following approximation:
          
         \deqn{kB-1 = (k Cd fc^2) / (4Ct ustar/u(zh)) + kBs-1(1 - fc)^2}
         
         If fc (fractional vegetation cover) is missing, it is estimated from LAI:

         \deqn{fc = 1 - exp(-LAI/2)}
         
         The wind speed at the top of the canopy is calculated using function
         \code{\link{wind.profile}}.
         
         Ct is the heat transfer coefficient of the leaf (Massman 1999):
         
         \deqn{Ct = Pr^-2/3 Reh^-1/2 N}
         
         where Pr is the Prandtl number (set to 0.71), and Reh is the Reynolds number for leaves:
         
         \deqn{Dl wind(zh) / v}
          
         kBs-1, the kB-1 value for bare soil surface, is calculated according 
         to Su et al. 2001:
         
         \deqn{kBs^-1 = 2.46(Re)^0.25 - ln(7.4)}
         
         Rb for water vapor and heat is assumed to be equal. Rb for CO2 (Rb_CO2) is given as:
 
         \deqn{Rb_CO2 = 1.4 * Rb}
 
         The factor 1.4 arises due the lower molecular diffusivity of CO2 compared to water.
         It is lower than the ratio of the molecular diffusivities (Dw/DCO2 = 1.6), as movement
         across the boundary layer is assumed to be partly by diffusion and partly by turbulent
         mixing (Nobel 2005).
}
\references{
Su, Z., Schmugge, T., Kustas, W. & Massman, W., 2001: An evaluation of
            two models for estimation of the roughness height for heat transfer between
            the land surface and the atmosphere. Journal of Applied Meteorology 40, 1933-1951.

            Massman, W., 1999: A model study of kB H- 1 for vegetated surfaces using
           'localized near-field' Lagrangian theory. Journal of Hydrology 223, 27-43.
           
            Nobel, P. S., 2005: Physicochemical and Environmental Plant Physiology. Third 
            Edition. Elsevier Academic Press, Burlington, USA.
}
\seealso{
\code{\link{Gb.Thom}}, \code{\link{Gb.Choudhury}}
}
