% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{intercellular.CO2}
\alias{intercellular.CO2}
\title{Bulk intercellular CO2 concentration}
\usage{
intercellular.CO2(Ca, GPP, Gs, surface.Ca = F, Ga = NULL, NEE = NULL,
  constants = bigleaf.constants())
}
\arguments{
\item{Ca}{Atmospheric CO2 concentration (umol mol-1)}

\item{GPP}{Gross primary productivity (umol CO2 m-2 s-1)}

\item{Gs}{Surface conductance to water (mol m-2 s-1)}

\item{surface.Ca}{Should the derived surface CO2 concentration be used instead of 
measured atmospheric CO2? If TRUE, Ca is derived as shown in \code{Details}.}

\item{Ga}{Aerodynamic conductance to CO2 (mol m-2 s-1)}

\item{NEE}{Net ecosystem exchange (umol CO2 m-2 s-1)}

\item{constants}{DwDc - Ratio of the molecular diffusivities for water vapor and CO2 (-)}
}
\value{
Ci - Bulk canopy intercellular CO2 concentration (umol mol-1)
}
\description{
Bulk canopy intercellular CO2 concentration (Ci) calculated based on Fick's law
             given surface conductance (Gs), gross primary productivity (GPP) and 
             atmospheric CO2 concentration (Ca).
}
\details{
Bulk intercellular CO2 concentration (Ci) is given as:

         \deqn{Ci = Ca - GPP/(Gs/1.6)}
         
         where Gs/1.6 (mol m-2 s-1) represents the surface conductance to CO2.
         Note that Gs is required in mol m-2 s-1.
}
\note{
The equation is based on Fick's law of diffusion and is equivalent to the
      often used equation at leaf level (ci = ca - An/gs).
      Note that GPP and Gs have a different interpretation than An and gs.
      Gs comprises non-pyhsiological contributions (i.e. physical evaporation)
      and is confounded by physical factors (e.g. energy balance non-closure).
      GPP does not account for dark respiration and is further subject to uncertainties
      in the NEE partitioning algorithm used.
      This function should be used with care and the resulting Ci might not be
      readily comparable to its leaf-level analogue.
}
