% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.r
\name{kg.to.mol}
\alias{kg.to.mol}
\title{Conversion between Mass and Molar Units}
\usage{
kg.to.mol(mass, molarMass = bigleaf.constants()$H2Omol)
}
\arguments{
\item{mass}{Numeric vector of mass in kg}

\item{molarMass}{Numeric vector of molar mass of the substance (kg mol-1)
e.g. as provided by \code{\link{bigleaf.constants}}()$H2Omol
Default is molar mass of Water.}
}
\value{
Numeric vector of amount of substance in mol.
}
\description{
Converts mass units of a substance to the corresponding molar units
             and vice versa.
}
