% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{LE.to.ET}
\alias{LE.to.ET}
\title{Conversion between LE and ET}
\usage{
LE.to.ET(LE, Tair)
}
\arguments{
\item{LE}{Latent heat flux (W m-2)}

\item{Tair}{Air temperature (deg C)}
}
\value{
ET Evapotranspiration (kg m-2 s-1)
}
\description{
converts water loss from mass (ET) to energy (ET) units, and vice versa
}

