% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{MoninObukhov.length}
\alias{MoninObukhov.length}
\title{Monin-Obukhov length}
\usage{
MoninObukhov.length(Tair, pressure, ustar, H, constants = bigleaf.constants())
}
\arguments{
\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{ustar}{Friction velocity (m s-1)}

\item{H}{Sensible heat flux (W m-2)}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
k - von Karman constant (-) \cr
g - gravitational acceleration (m s-2)}
}
\value{
Monin-Obukhov length (m)
}
\description{


The Monin-Obukhov length (L) is given by:

             \deqn{L = - (\rho * cp * ustar^3 * Tair) / (k * g * H)}
}
\note{
Note that L gets very small for very low ustar values with implications
      for subsequent functions using L as input. It is recommended to filter
      data and exclude low ustar values (ustar < ~0.2) beforehand.
}
\references{
Foken, T, 2008: Micrometeorology. Springer, Berlin, Germany.
}

