% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{ms.to.mol}
\alias{ms.to.mol}
\title{Conversion between conductance units}
\usage{
ms.to.mol(Gc_ms, Tair, pressure, constants)
}
\arguments{
\item{Gc_ms}{Canopy/Surface conductance (ms-1)}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (Pa)}

\item{constants}{}
}
\value{
Canopy/Surface conductance in mol m-2 s-1
}
\description{
Converts canopy/surface conductance from ms-1
             to mol m-2 s-1 (ms.to.mol), or vice versa (mol.to.ms) based on
             the ideal gas law.
}

