% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{ms.to.mol}
\alias{mol.to.ms}
\alias{ms.to.mol}
\title{Conversion between conductance units}
\usage{
ms.to.mol(G_ms, Tair, pressure, constants = bigleaf.constants())

mol.to.ms(G_mol, Tair, pressure, constants = bigleaf.constants())
}
\arguments{
\item{G_ms}{Conductance (ms-1)}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
Rgas - universal gas constant (J mol-1 K-1)}

\item{G_mol}{Conductance (mol m-2 s-1)}
}
\description{
Converts conductances from ms-1
             to mol m-2 s-1, or vice versa.
}
\references{
Jones, H.G. 1992. Plants and microclimate: a quantitative approach to environmental plant physiology.
            2nd Edition., 2nd Edn. Cambridge University Press, Cambridge. 428 p --> replace with third edition
}

