% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{pressure.from.elevation}
\alias{pressure.from.elevation}
\title{Air pressure from hypsometric equation}
\usage{
pressure.from.elevation(elev, Tair, q = NULL,
  constants = bigleaf.constants())
}
\arguments{
\item{elev}{Elevation a.s.l. (m)}

\item{Tair}{Air temperature (deg C)}

\item{q}{Specific humidity (kg kg-1); optional}

\item{constants}{Kelvin- conversion degC to Kelvin \cr
pressure0 - reference atmospheric pressure at sea level (Pa) \cr
Rd - gas constant of dry air (J kg-1 K-1) \cr
g -  gravitational acceleration (m s-2) \cr}
}
\value{
pressure air pressure (kPa)
}
\description{
An estimate of mean pressure at a given elevation as predicted by the
             hypsometric equation.
}
\details{

}
\note{
The hypsometric equation gives an estimate of the standard pressure
      at a given altitude. 
      If specific humidity q is provided, humidity correction
      is applied and the virtual temperature instead of air temperature is used.
}
\examples{
# mean pressure at 500m altitude at 25 deg C and specific humidity of 0.01 kg kg-1
pressure.from.elevation(500,Tair=25,q=0.01)

}
\references{
Stull B., 1988: An Introduction to Boundary Layer Meteorology.
            Kluwer Academic Publishers, Dordrecht, Netherlands.
}
