% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evapotranspiration.r
\name{reference.ET}
\alias{reference.ET}
\title{Reference Evapotranspiration}
\usage{
reference.ET(data, Gs_ref = 0.0143, Tair = "Tair", pressure = "pressure",
  VPD = "VPD", Rn = "Rn", Ga = "Ga", G = NULL, S = NULL,
  missing.G.as.NA = FALSE, missing.S.as.NA = FALSE,
  Esat.formula = c("Sonntag_1990", "Alduchov_1996", "Allen_1998"),
  constants = bigleaf.constants())
}
\description{
Reference evapotranspiration calculated from the Penman-Monteith
             equation with a prescribed surface conductance.
             This function is deprecated. Use potential.ET(...,approach="Penman-Monteith") instead
}
