% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanopyPhysics.R
\name{ReynoldsNumber}
\alias{ReynoldsNumber}
\title{Boundary layer conductance according to Thom 1972}
\usage{
ReynoldsNumber(hs, ustar, pressure, Tair, pressure0 = 101325,
  Tair0 = 273.15)
}
\arguments{
\item{ustar}{friction velocity (m s-1)}

\item{k}{von-Karman constant (-)}
}
\value{
a matrix containing Gb, Rb, and kB
}
\description{
An empirical formulation for the quasi-laminar boundary layer conductance
based on ustar
}

