% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{roughness.parameters}
\alias{roughness.parameters}
\title{Roughness parameters}
\usage{
roughness.parameters(method = c("wind_profile", "canopy_height"), zh,
  frac_d = 0.7, frac_z0m = 0.1, data, Tair = "Tair",
  pressure = "pressure", wind = "wind", ustar = "ustar", H = "H", zr,
  d = NULL, z0m = NULL, stab_roughness = TRUE,
  stab_formulation = c("Dyer_1970", "Businger_1971"),
  constants = bigleaf.constants())
}
\arguments{
\item{method}{Method used, either "canopy_height", or "wind_profile" \cr
NOTE: if method is "canopy_height", only the following three arguments
are used.}

\item{frac_d}{Fraction of displacement height on canopy height (-)}

\item{frac_z0m}{Fraction of roughness length on canopy height (-)
The following arguments are only needed if method = "wind_profile"!}

\item{data}{Data.frame or matrix containing all required variables}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{wind}{Wind speed at height zr (m s-1)}

\item{ustar}{Friction velocity (m s-1)}

\item{H}{Sensible heat flux (W m-2)}

\item{zr}{Instrument (reference) height (m)}

\item{d}{Zero-plane displacement height (m); optional}

\item{z0m}{Roughness length for momentum (m); optional}

\item{stab_roughness}{Should stability correction be considered? Default is TRUE}

\item{stab_formulation}{Stability correction function used (If stab_correction is TRUE).
Either "Dyer_1970" or "Businger_1971".}

\item{constants}{k - von-Karman constant (-) \cr
Kelvin - conversion degree Celsius to Kelvin \cr
cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
g - gravitational acceleration (m s-2) \cr}
}
\value{
a data.frame with the following columns:
        \item{d}{Zero-plane displacement height (m)}
        \item{z0m}{Roughness length for momentum (m)}


}
\description{
A simple approximation of the two roughness parameters displacement height (d)
             and roughness length for momentum (z0m).
}
\details{
The two main roughness parameters, the displacement height (d)
         and the roughness length for momentum (z0m) can be estimated from simple
         empirical relationships with canopy height (zh). If \code{method} is
         \code{canopy_height}, the following formulas are used:  
         
         \deqn{d = frac_d * zh}
         
         \deqn{z0m = frac_z0m * zh}
         
         where frac_d defaults to 0.7 and frac_z0m to 0.1.
         
         If \code{method} is \code{wind_profile}, z0m is estimated by solving
         the windspeed profile for z0m:
         
         \deqn{z0m = median((zr - d) * exp(-k*wind / ustar - psi_m)}
                 
         By default, d in this equation is fixed to 0.7zh, but can be set to any
         other value.
}

