% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{roughness.parameters}
\alias{roughness.parameters}
\title{Roughness parameters}
\usage{
roughness.parameters(method = c("canopy_height", "wind_profile"), zh,
  frac_d = 0.7, frac_z0m = 0.1, data, Tair = "Tair",
  pressure = "pressure", wind = "wind", ustar = "ustar", H = "H", zr,
  d = NULL, z0m = NULL, stab_roughness = T, formulation = c("Dyer_1970",
  "Businger_1971"), constants = bigleaf.constants())
}
\arguments{
\item{method}{method used, either "canopy_height", or "wind_profile" \cr
NOTE: if method is "canopy_height", only the following three arguments
are used.}

\item{frac_d}{fraction of displacement height on canopy height (-)}

\item{frac_z0m}{fraction of roughness length on canopy height (-)
The following arguments are only needed if method = "wind_profile"!}

\item{data}{a matrix or data.frame containing all required variables}

\item{Tair}{air temperature (deg C)}

\item{pressure}{air pressure (kPa)}

\item{wind}{wind speed at height zr (m s-1)}

\item{ustar}{friction velocity (m s-1)}

\item{H}{sensible heat flux (W m-2)}

\item{zr}{instrument height (m)}

\item{d}{optional; displacement height (m)}

\item{z0m}{optional; roughness length for momentum (m)}

\item{stab_roughness}{should stability correction be considered? Default is TRUE}

\item{formulation}{}

\item{constants}{k - von-Karman constant (-)
Kelvin - conversion degree Celsius to Kelvin
cp - specific heat of air for constant pressure (J K-1 kg-1)
g - gravitational acceleration (m s-2)}
}
\value{
a data.frame with the following columns:
        \item{d}{zero-plane displacement height (m)}
        \item{z0m}{roughness length for momentum (m)}


}
\description{
a simple approximation of the two roughness parameters displacement height (d)
             and roughness length for momentum (z0m).
}
\details{
The two main roughness parameters, the displacement height (d)
         and the roughness length for momentum (z0m) can be estimated from simple
         empirical relationships with canopy height (zh). If \code{method} is
         \code{canopy_height}, the following formulas are used:  
         
         \deqn{d = frac_d * zh}
         
         \deqn{z0m = frac_z0m * zh}
         
         where frac_d defaults to 0.7 and frac_z0m to 0.1.
         
         If \code{method} is \code{wind_profile}, z0m is estimated by solving
         the windspeed profile for z0m:
         
         \deqn{z0m = median((zr - d) * exp(-k*wind / ustar - psi_m)}
                 
         By default, d in this equation is fixed to 0.7zh, but can be set to any
         other value.
}

