% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{stability.correction}
\alias{stability.correction}
\title{stability correction functions for heat and momentum}
\usage{
stability.correction(data, Tair = "Tair", pressure = "pressure",
  ustar = "ustar", H = "H", zr, d, formulation = c("Dyer_1970",
  "Businger_1971"), constants = bigleaf.constants())
}
\arguments{
\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{ustar}{Friction velocity (m s-1)}

\item{H}{Sensible heat flux (W m-2)}

\item{zr}{instrument (reference) height (m)}

\item{d}{Zero-plane displacement height (m)}

\item{formulation}{Formulation for the stability function. Either "Dyer_1970", or "Businger_1971"}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
Rd - gas constant of dry air (J kg-1 K-1) \cr
cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
k - von Karman constant (-) \cr
g - gravitational acceleration (m s-2)}
}
\value{
a data.frame with the following columns:
         \item{psi_h}{the value of the stability function for heat and water vapor (-)}
         \item{psi_m}{the value of the stability function for momentum (-)}
}
\description{
dimensionless stability functions needed to correct deviations
             from the exponential wind profile under non-neutral conditions.
}
\details{
gives the integrated form of the universal functions.

The functions depend on the value of the stability parameter \eqn{\zeta},
         which can be calculated from the function \code{\link{stability.parameter}}.
         The integration of the universal functions is:
         
         \deqn{-x * zeta} 
         
         for stable atmospheric conditions (\eqn{\zeta} >= 0), and
         
         \deqn{2 * log( (1 + y) / 2) }
         
         for unstable atmospheric conditions (\eqn{\zeta} < 0).
         
         The different formulations differ in their value of x and y.
}
\references{
Dyer, A.J., 1974: A review of flux-profile relationships. 
            Boundary-Layer Meteorology 7, 363-372.
            
            Dyer, A. J., Hicks, B.B., 1970: Flux-Gradient relationships in the
            constant flux layer. Quart. J. R. Meteorol. Soc. 96, 715-721.
            
            Businger, J.A., Wyngaard, J. C., Izumi, I., Bradley, E. F., 1971:
            Flux-Profile relationships in the atmospheric surface layer. 
            J. Atmospheric Sci. 28, 181-189.
            
            Paulson, C.A., 1970: The mathematical representation of wind speed
            and temperature profiles in the unstable atmospheric surface layer.
            Journal of Applied Meteorology 9, 857-861.

            Foken, T, 2008: Micrometeorology. Springer, Berlin, Germany.
}

