% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{stability.parameter}
\alias{stability.parameter}
\title{stability parameter 'zeta'}
\usage{
stability.parameter(Tair, pressure, ustar, H, zr, d,
  constants = bigleaf.constants())
}
\arguments{
\item{Tair}{air temperature (degC)}

\item{pressure}{air pressure (kPa)}

\item{ustar}{friction velocity (m s-1)}

\item{H}{sensible heat flux (W m-2)}

\item{zr}{instrument height (m)}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
k - von Karman constant (-) \cr
g - gravitational acceleration (m s-2)}

\item{disp}{displacement height (m)}
}
\value{
the stability parameter \eqn{\zeta}
}
\description{
stability parameter 'zeta'
}
\details{
The stability parameter \eqn{\zeta} is given by:

         \deqn{\zeta = (zr - d) / L}
         
         where L is the Monin-Obukhov length, calculated from the function
         \code{\link{MoninObukhov.length}}. The displacement height d can 
         be estimated from the function \code{\link{roughness.parameters}}.
}

