% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{stomatal.slope}
\alias{stomatal.slope}
\title{Stomatal slope parameter "g1"}
\usage{
stomatal.slope(data, Tair = "Tair", pressure = "pressure", GPP = "GPP_nt",
  Gs = "Gs", VPD = "VPD", Ca = "Ca", model = c("USO", "Ball&Berry",
  "Leuning"), nmin = 40, fitg0 = FALSE, g0 = 0, fitD0 = FALSE,
  D0 = 1.5, Gamma = 50, constants = bigleaf.constants())
}
\arguments{
\item{data}{Data.frame or matrix containing all required columns}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Atmospheric pressure (kPa)}

\item{GPP}{Gross primary productivity (umol CO2 m-2 s-1)}

\item{Gs}{Surface conductance to water vapor (mol m-2 s-1)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{Ca}{Atmospheric CO2 concentration (umol mol-1)}

\item{model}{Stomatal model used. One of c("USO","Ball&Berry","Leuning")}

\item{nmin}{Minimum number of data required to perform the fit; defaults to 40.}

\item{fitg0}{Should g0 and g1 be fitted simultaneously?}

\item{g0}{Minimum stomatal conductance (mol m-2 s-1); ignored if \code{fitg0} is TRUE.}

\item{fitD0}{Should D0 be fitted along with g1 (and g0 if fitg0 = TRUE)?; only used if \code{model} is "Leuning"}

\item{D0}{Stomatal sensitivity parameter to VPD; only used if \code{model} is "Leuning" and fitD0 is FALSE}

\item{Gamma}{Canopy CO2 compensation point (umol mol-1); only used if \code{model} is "Leuning", defaults to 50 umol mol-1.}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin \cr
Rgas - universal gas constant (J mol-1 K-1)}
}
\value{
an nls model object, containing information on the fitted parameters, their uncertainty range,
        model fit, etc.
}
\description{
Estimation of the intrinsic WUE metric "g1" (stomatal slope) 
             from non-linear regression.
}
\details{
All stomatal models were developed at leaf-level, but its parameters 
         can also be estimated at ecosystem level (but be aware of caveats).
         
         The unified stomatal optimization (USO) model is given as (Medlyn et al. 2011):
     
         \deqn{gs = g0 + 1.6*(1.0 + g1/sqrt(VPD)) * GPP/Ca}
         
         The semi-empirical model by Ball et al. 1987 is defined as:
         
         \deqn{gs = g0 + g1* ((An * rH) / Ca)}
         
         Leuning 1995 suggested a revised version of the Ball&Berry model:
         
         \deqn{gs = g0 + g1*GPP / ((Ca - Gamma) * (1 + VPD/D0))}
         
         The parameters in the models are estimated using non-linear regression (\code{nls()}).
         Alternatively to measured VPD and Ca (i.e. conditions at instrument height), conditions at 
         the big-leaf surface can be provided. They can be calculated using \code{\link{surface.conditions}}.
}
\examples{
# calculate g1 for the site DE-Tha using data from June 2010
load("DE_Tha_2010") # load data

# filter data to ensure that Gs is a meaningful proxy to canopy conductance (Gc)
DE_Tha_2010 <- filter.data(DE_Tha_2010,quality.control=TRUE,quality.ext="_qc",
                           vars.qc=c("precip","Tair","VPD","GPP_nt","H","LE"),
                           tprecip=0.01,precip.hours=24,NA.as.precip=F,trad=200,
                           ttemp=5,tustar=0.2,tGPP=0.5,ws=15,min.int=5,trH=0.95) 

# select data only that were filtered in the previous function and in June
DE_Tha_June_2010 <- DE_Tha_2010[DE_Tha_2010[,"valid"] == 1 & DE_Tha_2010[,"month"] == 6,]

# calculate Gs from the the inverted PM equation (now Rn, and Ga are needed)
# calculate a simple estimate of Ga based on Thom 1972
Ga <- aerodynamic.conductance(DE_Tha_June_2010,stab_correction=F,
                              Rb_model="Thom_1972")[,"Ga_h"]

# if G and/or S are available, don't forget to indicate (they are ignored by default).
# Note that Ga is not added to the data.frame 'DE_Tha_June_2010'
Gs_PM <- surface.conductance(DE_Tha_June_2010,Tair="Tair",pressure="pressure",
                             Rn="Rn",G="G",S=NULL,VPD="VPD",Ga=Ga,PM=TRUE)[,"Gs_mol"]
                             
### Estimate the stomatal slope parameter g1 using the USO model
mod_USO <- stomatal.slope(DE_Tha_June_2010,model="USO",GPP="GPP_nt",Gs=Gs_PM,
                          nmin=40,fitg0=FALSE)

### Estimate the same parameter from the Ball&Berry model and prescribe g0
mod_BB <- stomatal.slope(DE_Tha_June_2010,model="Ball&Berry",GPP="GPP_nt",
                         Gs=Gs_PM,g0=0.01,nmin=40,fitg0=FALSE)                         

## same for the Leuning model, but this time estimate both g1 and g0 (but fix D0)
mod_Leu <- stomatal.slope(DE_Tha_June_2010,model="Leuning",GPP="GPP_nt",Gs=Gs_PM,
                          nmin=40,fitg0=TRUE,D0=1.5,fitD0=FALSE)                         

}
\references{
Medlyn, B.E., et al., 2011: Reconciling the optimal and empirical approaches to
            modelling stomatal conductance. Global Change Biology 17, 2134-2144.
            
            Ball, T.J., Woodrow, I.E., Berry, J.A. 1987: A model predicting stomatal conductance
            and its contribution to the control of photosynthesis under different environmental conditions.
            In: Progress in Photosynthesis Research, edited by J.Biggins, pp. 221-224, Martinus Nijhoff Publishers,
            Dordrecht, Netherlands.
            
            Leuning, R., 1995: A critical appraisal of a combined stomatal-photosynthesis
            model for C3 plants. Plant, Cell and Environment 18, 339-355.
}
