% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Temp.virtual}
\alias{Temp.virtual}
\title{virtual temperature}
\usage{
Temp.virtual(Tair, q)
}
\arguments{
\item{Tair}{air temperature (deg C)}

\item{q}{specific humidity (kg kg-1)}
}
\value{
virtual temperature (K)
}
\description{
virtual temperature
}
\details{
the virtual temperature is given by:
 \deqn{(Tair+273.15)(1 + 0.61*q)}
}
\note{
mixing ratio is approximated by specific humidity
}
\references{
Foken, T, 2008: Micrometeorology. Springer, Berlin, Germany.
}

