% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{Temp.virtual}
\alias{Temp.virtual}
\title{Virtual temperature}
\usage{
Temp.virtual(Tair, q, constants = bigleaf.constants())
}
\arguments{
\item{Tair}{Air temperature (deg C)}

\item{q}{Specific humidity (kg kg-1)}

\item{constants}{Kelvin - conversion degree Celsius to Kelvin}
}
\value{
virtual temperature (deg C)
}
\description{
Virtual temperature, defined as the temperature at which dry air would have the same
             density as moist air at its actual temperature.
}
\details{
the virtual temperature is given by:
 
 \deqn{Tv = (Tair+273.15)(1 + 0.61*q)}
}
\note{
mixing ratio is approximated by specific humidity.
}
\references{
Foken T., 2008: Micrometeorology. Springer, Berlin, Germany.
}

