% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{wetbulb.temp}
\alias{wetbulb.temp}
\title{Wet bulb temperature}
\usage{
wetbulb.temp(Tair, pressure, VPD, constants = bigleaf.constants())
}
\arguments{
\item{Tair}{Air temperature (deg C)}

\item{pressure}{Atmospheric pressure (kPa)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{constants}{cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
eps - ratio of the molecular weight of water vapor to dry air (-)}
}
\description{
calculates the wet bulb temperature, i.e. the temperature
             that the air would have if it was saturated.
}
\details{
Wet-bulb temperature (Tw) is calculated from the following expression:
         
         \deqn{e = Esat(Tw) - gamma* (Tair - Tw)}
         
         The equation is solved for Tw using \code{\link{nls}}.
         Actual vapor pressure e is calculated from VPD using the function \code{\link{VPD.to.e}}.
}
\examples{
wetbulb.temp(Tair=c(20,25),pressure=100,VPD=c(1,1.6))             
       
}
\references{
Monteith J.L., Unsworth M.H., 2008: Principles of Environmental Physics.
            3rd edition. Academic Press, London.
}
