% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_roughness.r
\name{wind.profile}
\alias{wind.profile}
\title{Wind Speed at a Given Height in the Surface Layer}
\usage{
wind.profile(
  data,
  z,
  Tair = "Tair",
  pressure = "pressure",
  ustar = "ustar",
  H = "H",
  wind = "wind",
  zr,
  zh,
  d = NULL,
  frac_d = 0.7,
  z0m = NULL,
  frac_z0m = NULL,
  estimate_z0m = TRUE,
  stab_correction = TRUE,
  stab_formulation = c("Dyer_1970", "Businger_1971"),
  constants = bigleaf.constants()
)
}
\arguments{
\item{data}{Data.frame or matrix containing all required variables}

\item{z}{Height above ground for which wind speed is calculated.}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Atmospheric pressure (kPa)}

\item{ustar}{Friction velocity (m s-1)}

\item{H}{Sensible heat flux (W m-2)}

\item{wind}{Wind speed at height zr (m s-1); only used if \code{stab_correction = TRUE}}

\item{zr}{Instrument (reference) height (m)}

\item{zh}{Canopy height (m)}

\item{d}{Zero-plane displacement height (-)}

\item{frac_d}{Fraction of displacement height on canopy height (-);
only used if \code{d} is not available}

\item{z0m}{Roughness length (m), optional; only used if \code{stab_correction = FALSE} (default=0.1)}

\item{frac_z0m}{Fraction of roughness length on canopy height (-), optional; only used if \code{z0m} is not provided.
Default is 0.1.}

\item{estimate_z0m}{Should \code{z0m} be estimated from the logarithmic wind profile? If \code{TRUE} (the default),
arguments \code{z0m} and \code{frac_z0m} are ignored.
See \code{\link{roughness.parameters}} for details.}

\item{stab_correction}{Should stability correction be applied? Defaults to \code{TRUE}}

\item{stab_formulation}{Stability correction function used (If \code{stab_correction = TRUE}).
Either \code{"Dyer_1970"} or \code{"Businger_1971"}.}

\item{constants}{k - von-Karman constant (-) \cr
Kelvin - conversion degree Celsius to Kelvin \cr
cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
g - gravitational acceleration (m s-2)}
}
\value{
A vector of wind speed at heights \code{z}.
}
\description{
Wind speed at a given height above the canopy estimated from single-level
             measurements of wind speed.
}
\details{
The underlying assumption is the existence of a logarithmic wind profile
         above the height d + z0m (the height at which wind speed mathematically reaches zero
         according to the Monin-Obukhov similarity theory).
         In this case, the wind speed at a given height z is given by:
         
           \deqn{u(z) = (ustar/k) * (ln((z - d) / z0m) - \psi{m}}
         
         The roughness parameters zero-plane displacement height (d) and roughness length (z0m)
         can be approximated from \code{\link{roughness.parameters}}. \eqn{\psi{m}} is omitted
         if \code{stab_correction = FALSE} (not recommended). If \code{estimate_z0m = TRUE},
         z0m is first estimated from the wind profile equation and then used in the equation
         above for the calculation of \code{u(z)} (see e.g. Newman & Klein 2014).
}
\note{
Note that this equation is only valid for z >= d + z0m, and it is not 
      meaningful to calculate values closely above d + z0m. All values in \code{heights}
      smaller than d + z0m will return 0.
}
\examples{
heights <- seq(18,40,2)  # heights above ground for which to calculate wind speed
df <- data.frame(Tair=25,pressure=100,wind=c(3,4,5),ustar=c(0.5,0.6,0.65),H=c(200,230,250)) 
ws <- sapply(heights,function(x) wind.profile(df,z=x,zr=40,zh=25,d=16))
colnames(ws) <- paste0(heights,"m")

}
\references{
Monteith, J.L., Unsworth, M.H., 2008: Principles of Environmental Physics.
            3rd edition. Academic Press, London. 
            
            Newman, J.F., Klein, P.M., 2014: The impacts of atmospheric stability on
            the accuracy of wind speed extrapolation methods. Resources 3, 81-105.
}
\seealso{
\code{\link{roughness.parameters}}
}
