% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{wind.profile}
\alias{wind.profile}
\title{Wind speed at given levels in the surface layer}
\usage{
wind.profile(heights, data, Tair = "Tair", pressure = "pressure",
  ustar = "ustar", H = "H", zr, d, z0m, stab_correction = TRUE,
  stab_formulation = c("Dyer_1970", "Businger_1971"),
  constants = bigleaf.constants())
}
\arguments{
\item{heights}{vector with heights for which wind speed is to be 
calculated.}

\item{data}{Data.frame or matrix containing all required variables}

\item{Tair}{Air temperature (deg C)}

\item{pressure}{Air pressure (kPa)}

\item{ustar}{Friction velocity (m s-1)}

\item{H}{Sensible heat flux (W m-2)}

\item{zr}{Instrument (reference) height (m)}

\item{d}{Zero-plane displacement height (m)}

\item{z0m}{Roughness length for momentum (m)}

\item{stab_correction}{Should stability correction be applied? Defaults to TRUE}

\item{stab_formulation}{Stability correction function used (If stab_correction is TRUE).
Either "Dyer_1970" or "Businger_1971".}

\item{constants}{k - von-Karman constant (-) \cr
Kelvin - conversion degree Celsius to Kelvin \cr
cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
g - gravitational acceleration (m s-2) \cr}
}
\value{
a data.frame with rows representing time and columns representing heights as specified in \code{heights}.
}
\description{
wind speed at a given height above the canopy estimated from single-level
             measurements of wind speed at some distance above the canopy.
}
\details{
The underlying assumption is the existence of a logarithmic wind profile
         above the height d + z0m (the height at which wind speed mathematically reaches zero
         according to the Monin-Obhukov similarity theory).
         In this case, the wind speed at a given height z is given by:
         
         \deqn{u(z) = (ustar/k) * (ln((z - d) / z0m) - \psi{m}}
         
         The roughness parameters zero-plane displacement height (d) and roughness length (z0m)
         can be approximated from \code{\link{roughness.parameters}}.
}
\note{
Note that this equation is only valid for z >= d + z0m. All values in \code{heights}
      smaller than d + z0m will return 0.
}
\references{

}

