% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{WUE.metrics}
\alias{WUE.metrics}
\title{Water Use Efficiency metrics}
\usage{
WUE.metrics(data, GPP = "GPP", NEE = "NEE", LE = "LE", VPD = "VPD",
  Tair = "Tair", constants = bigleaf.constants())
}
\arguments{
\item{data}{Data.frame or matrix containing all required columns}

\item{GPP}{Gross primary productivity (umol CO2 m-2 s-1)}

\item{NEE}{Net ecosystem exchange (umol CO2 m-2 s-1)}

\item{LE}{Latent heat flux (W m-2)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{Tair}{Air temperature (deg C)}

\item{constants}{}
}
\value{
a named vector with the following elements:
}
\description{
Calculation of various water use efficiency metrics
}
\details{
the following metrics are calculated:
}
\references{
Beer_2009
}

