% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigleaf.R
\name{WUE.metrics}
\alias{WUE.metrics}
\title{Water-Use Efficiency metrics}
\usage{
WUE.metrics(data, GPP = "GPP_nt", NEE = "NEE", LE = "LE", VPD = "VPD",
  Tair = "Tair", constants = bigleaf.constants())
}
\arguments{
\item{data}{data.frame or matrix containing all required variables}

\item{GPP}{Gross primary productivity (umol CO2 m-2 s-1)}

\item{NEE}{Net ecosystem exchange (umol CO2 m-2 s-1)}

\item{LE}{Latent heat flux (W m-2)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{Tair}{Air temperature (deg C)}

\item{constants}{Cmol - molar mass of carbon (kg mol-1)}
}
\value{
a named vector with the following elements:
        \item{WUE}{Water-use efficiency (gC (kg H20)-1)}
        \item{WUE_NEE}{Water-use efficiency based on NEE (gC (kg H20)-1)}
        \item{IWUE}{Inherent water-use efficiency (gC kPa (kg H20)-1)}
        \item{uWUE}{Underlying water-use efficiency (gC kPa (kg H20)-1)}
}
\description{
Calculation of various water use efficiency (WUE) metrics.
}
\details{
the following metrics are calculated:

         Water-use efficiency (WUE):
         
         \deqn{WUE = GPP / ET}
         
         Water-use efficiency based on NEE (WUE_NEE):
         
         \deqn{WUE_NEE = NEE / ET}
         
         Inherent water-use efficiency (IWUE; Beer et al. 2009):
         
         \deqn{IWUE = (GPP * VPD) / ET}
         
         Underlying water-use efficiency (uWUE; Zhou et al. 2014):
         
         \deqn{uWUE= (GPP * sqrt(VPD)) / ET}
         
         All metrics are calculated based on the median of all values. E.g.
         WUE = median(GPP/ET,na.rm=TRUE)
}
\note{
Units for VPD can also be hPa. Units change accordingly.
      WUE_NEE is calculated based on the absolute value of NEE (the sign convention does not matter).
}
\examples{
## filter data for dry periods and daytime at DE-Tha in June 2014
DE_Tha_Jun_2014 <- filter.data(DE_Tha_Jun_2014,quality.control=TRUE,filter.growseas=FALSE,
                               quality.ext="_qc",vars.qc=c("precip","Tair","VPD","H","LE"),
                               tprecip=0.01,precip.hours=24,NA.as.precip=FALSE,trad=200,
                               tustar=0.2,tGPP=0,ws=15,min.int=5)

## calculate WUE metrics in the selected periods
WUE.metrics(DE_Tha_Jun_2014[DE_Tha_Jun_2014[,"valid"] > 0,])
                        
}
\references{
Beer, C., et al., 2009: Temporal and among-site variability of inherent
            water use efficiency at the ecosystem level. Global Biogeochemical Cycles 23, GB2018.
            
            Zhou, S., et al., 2014: The effect of vapor pressure deficit on water
            use efficiency at the subdaily time scale. Geophysical Research Letters 41.
}
